Split RIDER core into focused modules
