+++ coolprompt/optimizer/rider/core/assistant.py
@@ -1,245 +1,57 @@
-"""RIDER RiderGenesis — elite автоматическая оптимизация промптов без размеченных данных.
+"""RIDER Genesis Ultra facade for CoolPrompt.
 
-Четыре режима с чёткой дифференциацией качества под разные ценовые tier'ы:
-
-  ┌──────────────┬──────┬──────┬─────────┬────────┐
-  │ Feature      │ light│ blitz│standard │ ultra+ │
-  ├──────────────┼──────┼──────┼─────────┼────────┤
-  │ Contract+routing│ ✓  │  ✓   │   ✓     │   ✓    │
-  │ Strategies    │  2  │  4   │   5     │  5+2   │
-  │ Preservation+repair│ ✓ │ ✓ │   ✓     │   ✓    │
-  │ GENESIS lessons│ —  │  ✓   │   ✓     │   ✓    │
-  │ FORGE memory  │  —  │  —   │   ✓     │   ✓    │
-  │ Constitutional audit│— │ ✓ │  ✓ (2x) │ ✓ (2x) │
-  │ Refine cycles │  —  │  1   │   2     │   2    │
-  │ Diversity-aware merge│—│ ✓ │   ✓     │   ✓    │
-  │ **Synthetic test eval**│— │— │ 3 tests │ 5 tests│
-  │ **Beam search**│ k=1 │ k=1 │  k=2    │  k=3   │
-  │ **Red-team adversarial**│—│—│   —    │   ✓    │
-  │ Bilingual adversarial│ —│—│   —     │ ✓ non-EN│
-  │ VORTEX on collapse│ —│ — │   —     │   ✓    │
-  │ API calls     │ ~5  │ ~11  │  ~22    │  ~33   │
-  │ Runtime       │ ~20s│ ~60s │ ~120s   │ ~180s  │
-  └──────────────┴──────┴──────┴─────────┴────────┘
-
-This production line folds full RIDER techniques into a data-free optimizer:
-v4 Ultra+ добавляет ещё 3 ключевых апгрейда для стандартных и ультра-режимов:
-
-  ▸ SYNTHETIC TEST-CASE EVAL — standard/ultra генерируют 3-5 синтетических test
-    inputs и ранжируют кандидатов по качеству РЕАЛЬНЫХ output'ов, а не по
-    эстетике prompt-текста. Это де-факто labeled fitness без разметки человека.
-  ▸ BEAM SEARCH top-k — standard (k=2), ultra (k=3) не схлопывают ветки рано.
-  ▸ RED-TEAM ADVERSARIAL — только ultra. Sonnet в роли атакующего ищет 3 edge
-    cases; при severity>=medium запускается hardening-refine с конкретными
-    fix-directives.
-  ▸ v4.1 ENUMERATION PRESERVATION — закрывает главный gap обнаруженный на
-    реальном батче 15 ЭКСМО-промптов. Детектит в оригинале:
-      * счётчик numbered items (если >=3) → draft не может потерять >40%
-      * Level/Step markers (Уровень N, Level N, Step N, ШАГ N, Tier N)
-      * category codes (1.1.а, 2.3.б, 3.2) → каждый должен остаться
-    При нарушении — repair call заставляет Sonnet восстановить структуру.
-
-This production line expands the data-free optimizer with full RIDER techniques
-(ITMO), при этом сохраняя бюджет вызовов light/blitz/standard/ultra.
-
-Ключевые улучшения v3 vs v2:
-  1. PROMPT CONTRACT EXTRACTION (вместо free-text _analyze_context):
-     Sonnet возвращает JSON с task_archetype / language / preservation_flags /
-     recommended_strategies / failure_modes / output_anchor. Контракт пробрасывается
-     во все последующие вызовы.
-  2. ADAPTIVE STRATEGY ROUTING (MARS CONFIG-lite):
-     Набор стратегий выбирается из контракта, а не фиксирован. Например,
-     prompt про отладку кода НЕ получит 'creative' стратегию.
-  3. GENESIS-LITE — lesson extraction:
-     Каждый pairwise compare вызов возвращает winner + WHY-объяснение. WHY
-     становится уроком (lesson), который пробрасывается в последующие
-     strategy/refine/merge вызовы как контекст успеха.
-  4. OPERATOR FORGE — per-strategy memory:
-     Лучшие outputs каждой стратегии запоминаются и инжектируются обратно,
-     когда стратегия вызывается повторно (как few-shot контекст).
-  5. PRESERVATION RULES + POST-HOC VALIDATOR:
-     Во все strategy/merge/refine промпты инжектируются preservation constraints
-     для code/placeholders/markdown/URLs. После генерации — dry check на
-     сохранность артефактов; при нарушении — 1 repair call.
-  6. CONSTITUTIONAL AUDIT (вместо generic _find_weaknesses):
-     Рубрика из 6 измерений (clarity, specificity, constraint_completeness,
-     output_anchoring, edge_case_coverage, brevity_vs_bloat). Возвращает top-3
-     failure dimensions + конкретные fix-directives.
-  7. DIVERSITY-AWARE MERGE (k-DPP-lite):
-     Перед merge'ом считается Jaccard-similarity между кандидатами. Если
-     top-2 слишком похожи (>0.85) — выбирается более разнообразная пара.
-     Если весь пул collapsed — Ultra переключается на VORTEX paradigm-shift.
-  8. VORTEX-ON-COLLAPSE (только Ultra):
-     При обнаружении semantic collapse в Phase 2 один из refinement слотов
-     тратится на VORTEX-style paradigm shift вместо рутинной мутации.
-
-Бюджет вызовов (не меняется по сравнению с v2):
-  - light     ~ 5 calls,   ~15s
-  - blitz     ~11 calls,   ~45s
-  - standard  ~19 calls,   ~70s
-  - ultra     ~25 calls,  ~120s
-
-Использование:
-    from rider.assistant import RiderGenesis
-
-    tuner = RiderGenesis()
-    tuner.run("Write a persuasive essay about climate change", mode="ultra")
-    print(tuner.final_prompt)
-    print(f"Improvement: {tuner.improvement:+.1f}%")
+The heavy implementation is split into focused mixins under
+``coolprompt.optimizer.rider.core`` so the public ``RiderGenesis`` class stays
+readable while preserving the Ultra optimizer behavior.
 """
 
 from __future__ import annotations
 
-import json
 import logging
 import os
-import re
-import time
-from concurrent.futures import ThreadPoolExecutor
-from typing import Any, Dict, List, Optional, Tuple, Type
+from typing import Any, Dict, List, Optional
 
-from pydantic import BaseModel, Field, field_validator
 from coolprompt.utils.prompt_templates.rider_templates import (
     RIDER_BATCH_RANK_PROMPT,
     RIDER_BILINGUAL_ADVERSARIAL_PROMPT,
     RIDER_COMPARE_PROMPT,
     RIDER_CONSTITUTIONAL_AUDIT_PROMPT,
-    RIDER_CONTRACT_PROMPT,
     RIDER_EVAL_RESPONSE_PROMPT,
     RIDER_MERGE_PROMPT,
     RIDER_PRESERVE_REPAIR_PROMPT,
     RIDER_QUALITY_PROMPT,
     RIDER_RED_TEAM_PROMPT,
     RIDER_REFINE_PROMPT,
-    RIDER_SAFE_CLASSIFICATION_PROMPT_EN,
-    RIDER_SAFE_CLASSIFICATION_PROMPT_RU,
-    RIDER_SAFE_CODE_PROMPT_EN,
-    RIDER_SAFE_CODE_PROMPT_RU,
-    RIDER_SAFE_GENERIC_PROMPT_EN,
-    RIDER_SAFE_GENERIC_PROMPT_RU,
-    RIDER_SAFE_TRANSLATION_PROMPT_EN,
-    RIDER_SAFE_TRANSLATION_PROMPT_RU,
     RIDER_STRATEGY_PROMPTS,
     RIDER_SYNTHETIC_TEST_PROMPT,
 )
 from rider.llm.client import LLMClient
 
+from .contract import RiderContractMixin
+from .memory import RiderMemoryMixin
+from .pipeline_config import RiderPipelineConfigMixin
+from .preservation import RiderPreservationMixin
+from .prompt_ops import RiderPromptOpsMixin
+from .run_modes import RiderRunModesMixin
+from .runtime import RiderRuntimeMixin
+from .synthetic_eval import RiderSyntheticEvalMixin
+from .ultra import RiderUltraMixin
+
 logger = logging.getLogger(__name__)
 
 
-# ---------------------------------------------------------------------------
-# Structured contract extraction
-# ---------------------------------------------------------------------------
-
-_CONTRACT_PROMPT = RIDER_CONTRACT_PROMPT
-
-
-class _PromptContractSchema(BaseModel):
-    """Instructor/Pydantic schema for the RIDER Genesis task contract."""
-
-    task_archetype: str = "other"
-    language: str = "unknown"
-    domain: str = "general"
-    audience: str = "general"
-    output_format_anchor: str = "free-form"
-    must_preserve: List[str] = Field(default_factory=list)
-    failure_modes: List[str] = Field(default_factory=list)
-    recommended_strategies: List[str] = Field(default_factory=list)
-    avoid_strategies: List[str] = Field(default_factory=list)
-
-    @field_validator("must_preserve", "failure_modes", "recommended_strategies", "avoid_strategies", mode="before")
-    @classmethod
-    def _coerce_string_list(cls, value: Any) -> List[str]:
-        if value is None:
-            return []
-        if isinstance(value, str):
-            return [value]
-        if isinstance(value, list):
-            return [str(item) for item in value if str(item).strip()]
-        return [str(value)]
-
-
-class _SyntheticTestsSchema(BaseModel):
-    """Instructor schema for synthetic evaluation tests."""
-
-    tests: List[str] = Field(default_factory=list, min_length=1)
-
-    @field_validator("tests", mode="before")
-    @classmethod
-    def _coerce_tests(cls, value: Any) -> List[str]:
-        if isinstance(value, dict):
-            value = value.get("tests") or value.get("items") or value.get("inputs") or []
-        if isinstance(value, str):
-            return [value]
-        if not isinstance(value, list):
-            return []
-        tests: List[str] = []
-        for item in value:
-            if isinstance(item, dict):
-                item = item.get("input") or item.get("text") or item.get("case") or item
-            text = item if isinstance(item, str) else json.dumps(item, ensure_ascii=False)
-            if text.strip():
-                tests.append(text.strip())
-        return tests
-
-
-class _RedTeamAdversarialSchema(BaseModel):
-    """Instructor schema for Ultra red-team findings."""
-
-    edge_cases: List[str] = Field(default_factory=list)
-    severity: str = "low"
-    fix_directives: List[str] = Field(default_factory=list)
-
-    @field_validator("edge_cases", "fix_directives", mode="before")
-    @classmethod
-    def _coerce_items(cls, value: Any) -> List[str]:
-        if value is None:
-            return []
-        if isinstance(value, str):
-            return [value]
-        if isinstance(value, list):
-            return [str(item) for item in value if str(item).strip()]
-        return [str(value)]
-
-
-class _JudgeScoreSchema(BaseModel):
-    """Instructor schema for synthetic judge scores."""
-
-    score: int = Field(ge=1, le=10)
-
-    @field_validator("score", mode="before")
-    @classmethod
-    def _coerce_score(cls, value: Any) -> int:
-        if isinstance(value, str):
-            match = re.search(r"\b(10|[1-9])\b", value)
-            if match:
-                return int(match.group(1))
-        return int(value)
-
-
-# ---------------------------------------------------------------------------
-# Main class
-# ---------------------------------------------------------------------------
-
-
-class RiderGenesis:
-    """Elite automatic prompt optimization without labeled data.
-
-    Pipeline v3 (Multi-Strategy Prompt Enhancement + RIDER-inspired extras):
-      1. Prompt CONTRACT extraction (structured JSON via Sonnet)
-      2. ADAPTIVE strategy routing (MARS CONFIG-lite, from contract)
-      3. Parallel strategies with OPERATOR FORGE memory and injected preservation rules
-      4. Pairwise compare with GENESIS-lite lessons (winner + WHY)
-      5. Merge top-2 (diversity-aware, VORTEX on collapse for Ultra)
-      6. CONSTITUTIONAL audit (6-dim rubric) + refine
-      7. Preservation POST-HOC validator + repair (conditional)
-
-    Args:
-        model: optional OpenRouter worker-model override; by default the mode
-            selects a latest-model role ensemble.
-        api_key: OpenRouter API key (default: from OPENROUTER_API_KEY)
-        verbose: print progress
-    """
+class RiderGenesis(
+    RiderRuntimeMixin,
+    RiderMemoryMixin,
+    RiderContractMixin,
+    RiderPipelineConfigMixin,
+    RiderPromptOpsMixin,
+    RiderPreservationMixin,
+    RiderRunModesMixin,
+    RiderUltraMixin,
+    RiderSyntheticEvalMixin,
+):
+    """Elite automatic prompt optimization without labeled data."""
 
     # Backward-compatible constant for older imports. Real runs use the role
     # maps below so worker/planner/judge/critic can be different models.
@@ -372,2812 +184,6 @@ class RiderGenesis:
         # v4 cross-prompt persistent lesson cache.
         self._lesson_cache: Dict[str, List[str]] = self._load_lesson_cache()
 
-    @staticmethod
-    def _split_model_chain(value: Optional[str]) -> List[str]:
-        if not value:
-            return []
-        return [m.strip() for m in value.split(",") if m.strip()]
-
-    @classmethod
-    def _env_model_chain(cls, mode: str, role: str) -> List[str]:
-        keys = (
-            f"RIDER_GENESIS_{mode.upper()}_{role.upper()}_MODELS",
-            f"RIDER_GENESIS_{mode.upper()}_{role.upper()}_MODEL",
-            f"RIDER_GENESIS_{role.upper()}_MODELS",
-            f"RIDER_GENESIS_{role.upper()}_MODEL",
-        )
-        for key in keys:
-            chain = cls._split_model_chain(os.environ.get(key))
-            if chain:
-                return chain
-        return []
-
-    @classmethod
-    def _dedupe_models(cls, models: List[str], allow_blocked: bool = False) -> List[str]:
-        seen = set()
-        result: List[str] = []
-        for model in models:
-            if not allow_blocked and any(model.startswith(prefix) for prefix in cls._BLOCKED_MODEL_PREFIXES):
-                continue
-            if model and model not in seen:
-                result.append(model)
-                seen.add(model)
-        return result
-
-    @staticmethod
-    def _allow_chinese_model_fallbacks() -> bool:
-        return os.environ.get("RIDER_GENESIS_ALLOW_CHINESE_MODELS", "").strip().lower() in {
-            "1", "true", "yes", "on",
-        }
-
-    def _record_llm_attempt(
-        self,
-        *,
-        role: str,
-        model: str,
-        status: str,
-        reason: str = "",
-        structured: bool = False,
-    ) -> None:
-        self._llm_attempts.append({
-            "role": role,
-            "model": model,
-            "status": status,
-            "reason": reason,
-            "structured": structured,
-        })
-
-    def _build_role_model_chains(self, mode: str) -> Dict[str, List[str]]:
-        base = self._MODE_ROLE_MODELS.get(mode, self._MODE_ROLE_MODELS["standard"])
-        chains: Dict[str, List[str]] = {}
-        for role in self._ROLES:
-            chain = self._env_model_chain(mode, role) or list(base.get(role, []))
-            if role == "worker" and self._model_override:
-                chain = [self._model_override] + chain
-            if self.FALLBACK_MODEL not in chain:
-                chain.append(self.FALLBACK_MODEL)
-            chains[role] = self._dedupe_models(chain)
-        return chains
-
-    def _role_models(self, role: str = "worker") -> List[str]:
-        if not self._role_model_chains:
-            self._role_model_chains = self._build_role_model_chains(self._mode or "standard")
-        chain = self._role_model_chains.get(role) or self._role_model_chains.get("worker") or [self.model]
-        return chain
-
-    def _role_model(self, role: str = "worker") -> str:
-        return self._role_models(role)[0]
-
-    def _strategy_role(self, strategy: str) -> str:
-        # Analytical / techniques / VORTEX variants are critic-style mutations:
-        # they first look for failure modes, then rewrite. Other strategies are
-        # worker/synthesis mutations.
-        if strategy in {"analytical", "techniques", "vortex"}:
-            return "critic"
-        return "worker"
-
-    @classmethod
-    def _load_lesson_cache(cls) -> Dict[str, List[str]]:
-        try:
-            if os.path.exists(cls._LESSON_CACHE_PATH):
-                with open(cls._LESSON_CACHE_PATH, 'r', encoding='utf-8') as f:
-                    data = json.load(f)
-                if isinstance(data, dict):
-                    return {k: [str(x) for x in v][:cls._LESSON_CACHE_MAX_PER_KEY]
-                            for k, v in data.items() if isinstance(v, list)}
-        except Exception:
-            pass
-        return {}
-
-    def _save_lesson_cache(self):
-        try:
-            os.makedirs(os.path.dirname(self._LESSON_CACHE_PATH), exist_ok=True)
-        except Exception:
-            pass
-        try:
-            with open(self._LESSON_CACHE_PATH, 'w', encoding='utf-8') as f:
-                json.dump(self._lesson_cache, f, ensure_ascii=False, indent=2)
-        except Exception:
-            pass
-
-    def _cache_key(self) -> str:
-        archetype = self._contract.get('task_archetype', 'other') or 'other'
-        domain = self._contract.get('domain', 'general') or 'general'
-        # Normalize to lowercase, first word of domain only (to merge related domains).
-        d = str(domain).lower().split()[0] if str(domain).strip() else 'general'
-        return f"{archetype}::{d}"
-
-    def _prefetch_cached_lessons(self) -> int:
-        """Inject cached lessons from prior runs for the same archetype+domain.
-        Standard/Ultra use this; light/blitz don't (by design)."""
-        if self._mode not in ('standard', 'ultra'):
-            return 0
-        key = self._cache_key()
-        cached = self._lesson_cache.get(key) or []
-        if not cached:
-            return 0
-        # Dedup against current run lessons.
-        existing = {str(x)[:120] for x in self._lessons}
-        injected = 0
-        for lsn in cached[:8]:
-            if str(lsn)[:120] not in existing:
-                self._lessons.append(lsn)
-                existing.add(str(lsn)[:120])
-                injected += 1
-        return injected
-
-    def _persist_new_lessons(self):
-        """After run, merge this run's lessons into the disk cache (standard/ultra only)."""
-        if self._mode not in ('standard', 'ultra'):
-            return
-        if not self._lessons:
-            return
-        key = self._cache_key()
-        existing = list(self._lesson_cache.get(key) or [])
-        existing_set = {str(x)[:120] for x in existing}
-        for lsn in self._lessons:
-            sig = str(lsn)[:120]
-            if sig not in existing_set:
-                existing.append(lsn)
-                existing_set.add(sig)
-        # Keep only the most recent N per key.
-        self._lesson_cache[key] = existing[-self._LESSON_CACHE_MAX_PER_KEY:]
-        self._save_lesson_cache()
-
-    # ══════════════════════════════════════════════════════════════════════
-    # Contract extraction + adaptive routing
-    # ══════════════════════════════════════════════════════════════════════
-
-    _VALID_STRATEGIES = ('structural', 'analytical', 'creative', 'depth', 'techniques')
-
-    _DEFAULT_CONTRACT: Dict[str, Any] = {
-        'task_archetype': 'other',
-        'language': 'unknown',
-        'domain': 'general',
-        'audience': 'general audience',
-        'output_format_anchor': 'free-form',
-        'must_preserve': [],
-        'failure_modes': [],
-        'recommended_strategies': ['structural', 'analytical', 'creative', 'depth', 'techniques'],
-        'avoid_strategies': [],
-        'domain_rules': [],
-        'quality_dimensions': [],
-    }
-
-    _MODERATION_HINTS = (
-        'moderation', 'censor', 'classification', 'json array', 'risk', 'level',
-        'модерац', 'цензор', 'классифик', 'запрещ', 'нарушен', 'уровень',
-        'экстремизм', 'терроризм', 'наркотик', 'пропаганд', 'дискредитац',
-        'госбезопас', 'ненавист', 'вражд', 'нацизм', 'порнограф',
-        'модерац', 'цензор', 'классифик', 'запрещ', 'нарушен', 'уровень',
-        'экстремизм', 'терроризм', 'наркотик', 'пропаганд', 'дискредитац',
-        'госбезопас', 'ненавист', 'вражд', 'нацизм', 'порнограф',
-    )
-    _TRANSLATION_HINTS = (
-        'translation', 'translator', 'translate', '<translation>', '<dictionary>',
-        'перевод', 'переводчик', '<перевод>', 'словар',
-        'оригинал', 'русский текст', 'английского на русский',
-        'перевод', 'переводчик', '<перевод>', '<dictionary>', 'словар',
-        'оригинал', 'русский текст', 'английского на русский',
-    )
-    _GEO_RISK_HINTS = (
-        'россия', 'росси', 'украина', 'украин', 'рф', 'российск', 'украинск', 'вооруженных сил',
-        'вооружённых сил', 'дискредитац', 'антигосударствен', 'госбезопас',
-        'сво', 'армии',
-        'россия', 'украина', 'рф', 'российск', 'украинск', 'вооруженных сил',
-        'вооружённых сил', 'дискредитац', 'антигосударствен', 'госбезопас',
-        'сво', 'армии',
-    )
-
-    @staticmethod
-    def _append_unique(items: List[Any], value: str) -> None:
-        value = str(value).strip()
-        if value and value not in items:
-            items.append(value)
-
-    @staticmethod
-    def _contains_any(text_lower: str, needles: Tuple[str, ...]) -> bool:
-        return any(n in text_lower for n in needles)
-
-    def _enrich_contract_with_static_analysis(
-        self, prompt: str, contract: Dict[str, Any],
-    ) -> Dict[str, Any]:
-        """Deterministically enrich LLM contract for production prompts.
-
-        The LLM contract is good at broad intent. This layer catches hard production
-        invariants that must never depend on taste: JSON schemas, legal level scales,
-        category labels, dictionary tags, and risk-calibration rules.
-        """
-        c = dict(self._DEFAULT_CONTRACT)
-        c.update({k: v for k, v in contract.items() if k in self._DEFAULT_CONTRACT})
-
-        must = list(c.get('must_preserve') or [])
-        failures = list(c.get('failure_modes') or [])
-        rules = list(c.get('domain_rules') or [])
-        dims = list(c.get('quality_dimensions') or [])
-        avoid = [s for s in (c.get('avoid_strategies') or []) if s in self._VALID_STRATEGIES]
-
-        prompt_lower = (prompt or '').lower()
-        archetype = str(c.get('task_archetype') or '').lower()
-        is_moderation = self._contains_any(prompt_lower, self._MODERATION_HINTS)
-        is_translation = self._contains_any(prompt_lower, self._TRANSLATION_HINTS)
-        is_extraction = archetype == 'data_extraction' or self._contains_any(
-            prompt_lower,
-            (
-                'ocr', 'рукопис', 'расшифр', 'транскрип', 'наборщик', 'оглавлен',
-                'table of contents', 'toc', 'извлеч', 'extract',
-            ),
-        )
-        has_json_schema = bool(re.search(r'"(?:quote|context|risk|categories|level|translation|reason)"', prompt or ''))
-        has_levels = bool(re.search(r'\b(?:level|уровень)[ \t]*[0-3]\b', prompt or '', re.IGNORECASE))
-        has_geo_risk = self._contains_any(prompt_lower, self._GEO_RISK_HINTS)
-        is_code = archetype in {'code_generation', 'code_review', 'debugging'} or bool(
-            re.search(r"```|def |class |function |typescript|python|javascript|sql|api|stack trace|traceback", prompt_lower)
-        )
-
-        if is_moderation or has_json_schema or has_levels:
-            c['task_archetype'] = 'classification'
-            if c.get('domain') in (None, '', 'general'):
-                c['domain'] = 'content moderation / legal compliance'
-            if c.get('output_format_anchor') in (None, '', 'free-form'):
-                c['output_format_anchor'] = 'strict machine-parseable JSON'
-
-            for item in (
-                'exact JSON output schema and allowed field names',
-                'category labels / legal taxonomy names',
-                'level scale and calibration thresholds',
-                'safe-context filters and false-positive/false-negative policy',
-                'quote/context/location copying rules',
-                'legal references and article numbers',
-            ):
-                self._append_unique(must, item)
-            for failure in (
-                'schema drift or extra prose outside JSON',
-                'level calibration drift',
-                'exclusive-output contract drift',
-                'priority order drift',
-                'level cap or exception drift',
-                'false negatives on level 2/3 risk',
-                'generic risk text not grounded in quote',
-                'invented legal category or article',
-            ):
-                self._append_unique(failures, failure)
-            for rule in (
-                'For moderation/legal prompts, preserve the original legal taxonomy, output schema, and level scale before improving wording.',
-                'Prefer bounded legal precision over creative reframing; never invent categories, laws, fields, or thresholds not present in the source prompt.',
-                'False negatives on high-risk items are more severe than conservative inclusion, but explicit safe-context filters must remain intact.',
-                'Every risk explanation should cite a concrete signal from quote/input text, not a vague topic-level suspicion.',
-                'If the source prompt says there is only one allowed output, keep that output contract exclusive: no prose, no alternative formats, no helper text.',
-                'Preserve priority order, escalation rules, caps, and exceptions exactly before adding any clarifying wording.',
-            ):
-                self._append_unique(rules, rule)
-            for dim in (
-                'schema_compliance',
-                'level_calibration',
-                'exclusive_output_compliance',
-                'priority_order_preservation',
-                'level_cap_semantics',
-                'false_negative_resistance',
-                'false_positive_control',
-                'quote_grounding',
-            ):
-                self._append_unique(dims, dim)
-            if 'creative' not in avoid:
-                avoid.append('creative')
-            c['recommended_strategies'] = [
-                s for s in ('analytical', 'structural', 'techniques', 'depth', 'creative')
-                if s not in avoid
-            ]
-
-        if has_geo_risk:
-            for item in (
-                'Russia/Ukraine/Russian-side geopolitical escalation rules',
-                'maximum level caps and Level 0 exclusion rules',
-            ):
-                self._append_unique(must, item)
-            for failure in (
-                'dropping Russia/Ukraine escalation policy',
-                'raising Level 0 items when the source prompt forbids it',
-                'raising above Level 3',
-            ):
-                self._append_unique(failures, failure)
-            self._append_unique(
-                rules,
-                'If the source prompt contains Russia/Ukraine/Russian-side escalation logic, preserve its exact cap semantics: raise only as specified, never above Level 3, and do not escalate Level 0 unless the source explicitly says so.',
-            )
-            for dim in ('geopolitical_escalation_semantics', 'level_cap_semantics'):
-                self._append_unique(dims, dim)
-
-        if is_translation:
-            c['task_archetype'] = 'translation'
-            if c.get('domain') in (None, '', 'general'):
-                c['domain'] = 'literary translation'
-            if c.get('output_format_anchor') in (None, '', 'free-form'):
-                c['output_format_anchor'] = 'translation output specified by source prompt'
-            for item in (
-                'dictionary/glossary tags and term-handling policy',
-                'translation-only or wrapper-tag output format',
-                'source-language to Russian literary style contract',
-                'author rhythm, syntax, register, and intentional roughness',
-            ):
-                self._append_unique(must, item)
-            for failure in (
-                'over-smoothing author syntax and rhythm',
-                'ignoring dictionary terms or names',
-                'adding commentary when output must be translation-only',
-                'forcing verse or genre conventions not requested by source',
-            ):
-                self._append_unique(failures, failure)
-            for rule in (
-                'For literary translation prompts, preserve the author-specific rhythm and syntax instead of normalizing everything into generic smooth Russian prose.',
-                'Dictionary/glossary instructions are hard constraints; if they conflict with factual or stylistic fidelity, the prompt must specify how to resolve the conflict.',
-                'Do not add pre-translation analysis to user-visible output unless the source prompt explicitly requires it.',
-            ):
-                self._append_unique(rules, rule)
-            for dim in (
-                'fidelity_to_source',
-                'authorial_rhythm',
-                'glossary_compliance',
-                'no_extra_commentary',
-            ):
-                self._append_unique(dims, dim)
-            # Literary translation can use creative language, but only after structural
-            # and analytical constraints have locked the production contract.
-            c['recommended_strategies'] = [
-                s for s in ('analytical', 'structural', 'depth', 'techniques', 'creative')
-                if s not in avoid
-            ]
-
-        if is_code:
-            for item in (
-                'code blocks, inline code, filenames, API names, and placeholders',
-                'runtime constraints, error messages, stack traces, and expected behavior',
-                'public interfaces and backward-compatibility requirements',
-            ):
-                self._append_unique(must, item)
-            for failure in (
-                'renaming public API or config keys',
-                'dropping error context or reproduction steps',
-                'inventing dependencies not requested',
-                'changing behavior outside the requested scope',
-            ):
-                self._append_unique(failures, failure)
-            for rule in (
-                'For code prompts, preserve exact identifiers, file paths, stack traces, and public contracts; improve debugging structure without fictional APIs.',
-                'Prefer minimal, testable fixes and explicit verification steps over creative rewrites.',
-            ):
-                self._append_unique(rules, rule)
-            for dim in ('identifier_preservation', 'testability', 'scope_control', 'runtime_correctness'):
-                self._append_unique(dims, dim)
-            if 'creative' not in avoid:
-                avoid.append('creative')
-            c['recommended_strategies'] = [
-                s for s in ('analytical', 'structural', 'techniques', 'depth', 'creative')
-                if s not in avoid
-            ]
-
-        if is_extraction and not is_moderation:
-            for item in (
-                'verbatim source text / source structure',
-                'markup tags, placeholders, page numbers, and line breaks',
-                'no commentary outside the requested extraction output',
-            ):
-                self._append_unique(must, item)
-            for failure in (
-                'paraphrasing instead of extraction',
-                'omitting source lines or structural elements',
-                'adding explanations around extracted output',
-                'inventing unreadable or missing text',
-            ):
-                self._append_unique(failures, failure)
-            for rule in (
-                'For extraction/OCR prompts, optimize for exact source preservation and completeness, not richer prose.',
-                'Do not add creative roleplay, examples, or extra output sections if they make the extraction contract less literal.',
-            ):
-                self._append_unique(rules, rule)
-            for dim in ('verbatim_fidelity', 'completeness', 'markup_compliance', 'no_extra_commentary'):
-                self._append_unique(dims, dim)
-            if 'creative' not in avoid:
-                avoid.append('creative')
-            c['recommended_strategies'] = [
-                s for s in ('analytical', 'structural', 'techniques', 'depth', 'creative')
-                if s not in avoid
-            ]
-
-        if archetype in {'creative_writing', 'analytical_essay', 'persuasion', 'brainstorming'}:
-            for item in (
-                'original user topic, audience, tone, and requested genre',
-                'the requested output type (essay, story, argument, idea list, etc.)',
-            ):
-                self._append_unique(must, item)
-            for failure in (
-                'generic classroom prompt with no distinctive angle',
-                'overly broad structure that produces predictable writing',
-                'missing sensory/detail/examples requirements',
-                'tone mismatch with the requested genre',
-            ):
-                self._append_unique(failures, failure)
-            for rule in (
-                'For short creative or essay prompts, produce a full high-leverage instruction: role, concrete angle, structure, sensory/detail requirements, quality bar, and anti-generic constraints.',
-                'Make the prompt immediately usable by a strong model; avoid vague meta-advice and ask for specific rhetorical moves.',
-            ):
-                self._append_unique(rules, rule)
-            for dim in ('distinctive_angle', 'structure_quality', 'specificity', 'anti_genericity'):
-                self._append_unique(dims, dim)
-            c['recommended_strategies'] = [
-                s for s in ('creative', 'structural', 'depth', 'analytical', 'techniques')
-                if s not in avoid
-            ]
-
-        c['must_preserve'] = must[:40]
-        c['failure_modes'] = failures[:24]
-        c['domain_rules'] = rules[:20]
-        c['quality_dimensions'] = dims[:16]
-        c['avoid_strategies'] = avoid
-        return c
-
-    def _extract_contract(self, prompt: str) -> Dict[str, Any]:
-        """Extract structured prompt contract via Instructor/Pydantic. Falls back to defaults."""
-        meta = _CONTRACT_PROMPT.format(prompt=prompt)
-        try:
-            obj = self._generate_structured(
-                prompt=meta,
-                schema=_PromptContractSchema,
-                role="planner",
-                temperature=0.2, max_tokens=600,
-                allowed_starts=("{",),
-                max_retries=2,
-            )
-            if obj is None:
-                raise ValueError("structured contract unavailable")
-            data = self._schema_to_dict(obj)
-            # Merge with defaults so missing fields don't break downstream code.
-            contract = dict(self._DEFAULT_CONTRACT)
-            contract.update({k: v for k, v in data.items() if k in self._DEFAULT_CONTRACT})
-            # Sanitize strategy lists.
-            rec = [s for s in contract.get('recommended_strategies', []) if s in self._VALID_STRATEGIES]
-            if not rec:
-                rec = list(self._VALID_STRATEGIES)
-            avoid = [s for s in contract.get('avoid_strategies', []) if s in self._VALID_STRATEGIES]
-            rec = [s for s in rec if s not in avoid]
-            if not rec:
-                rec = [s for s in self._VALID_STRATEGIES if s not in avoid] or list(self._VALID_STRATEGIES)
-            contract['recommended_strategies'] = rec
-            contract['avoid_strategies'] = avoid
-            if not isinstance(contract.get('must_preserve'), list):
-                contract['must_preserve'] = []
-            if not isinstance(contract.get('failure_modes'), list):
-                contract['failure_modes'] = []
-            if not isinstance(contract.get('domain_rules'), list):
-                contract['domain_rules'] = []
-            if not isinstance(contract.get('quality_dimensions'), list):
-                contract['quality_dimensions'] = []
-            return self._enrich_contract_with_static_analysis(prompt, contract)
-        except Exception as exc:
-            logger.info(f"Contract extraction failed, using defaults: {exc}")
-            return self._enrich_contract_with_static_analysis(prompt, dict(self._DEFAULT_CONTRACT))
-
-    @staticmethod
-    def _extract_json_value(text: str, allowed_starts: Tuple[str, ...] = ('{', '[')) -> Optional[str]:
-        """Extract the first valid JSON object/array text from LLM output."""
-        if not text:
-            return None
-        # Strip common code fences.
-        text = re.sub(r"```(?:json)?\s*", "", text.strip())
-        text = text.replace("```", "").strip()
-        decoder = json.JSONDecoder()
-        for start, ch in enumerate(text):
-            if ch not in allowed_starts:
-                continue
-            try:
-                _value, end = decoder.raw_decode(text[start:])
-            except json.JSONDecodeError:
-                continue
-            return text[start:start + end]
-        return None
-
-    @staticmethod
-    def _extract_json(text: str) -> Optional[str]:
-        """Extract the first top-level {...} JSON object from LLM output."""
-        return RiderGenesis._extract_json_value(text, allowed_starts=('{',))
-
-    def _adaptive_strategies(self, mode: str) -> List[str]:
-        """Pick the strategy set for a mode based on contract.recommended_strategies."""
-        rec = list(self._contract.get('recommended_strategies', self._VALID_STRATEGIES))
-        avoid = set(self._contract.get('avoid_strategies', []))
-
-        # Ordering preference comes from contract.
-        ordered = [s for s in rec if s in self._VALID_STRATEGIES and s not in avoid]
-        # Fallback: fill with unused valid strategies in canonical order.
-        for s in self._VALID_STRATEGIES:
-            if s not in ordered and s not in avoid:
-                ordered.append(s)
-        # Hard floor in case the contract zeroes everything out.
-        if not ordered:
-            ordered = list(self._VALID_STRATEGIES)
-
-        mode_budget = {'light': 2, 'blitz': 4, 'standard': 5, 'ultra': 5}.get(mode, 5)
-        selected = ordered[:mode_budget]
-
-        # Light always keeps analytical+structural in the mix for robustness, but lets
-        # the contract reorder: if contract prefers 'depth, techniques' for a code prompt,
-        # respect it instead of forcing analytical.
-        return selected
-
-    def _strategy_allowed(self, strategy: str) -> bool:
-        return strategy in self._VALID_STRATEGIES and strategy not in set(self._contract.get('avoid_strategies') or [])
-
-    def _mutation_strategy(self, preferred: str, fallback: str = 'techniques') -> str:
-        if self._strategy_allowed(preferred):
-            return preferred
-        if self._strategy_allowed(fallback):
-            return fallback
-        for strategy in ('analytical', 'structural', 'depth', 'techniques', 'creative'):
-            if self._strategy_allowed(strategy):
-                return strategy
-        return fallback
-
-    # ══════════════════════════════════════════════════════════════════════
-    # Context block builders (injected into strategy/compare/refine prompts)
-    # ══════════════════════════════════════════════════════════════════════
-
-    def _contract_block(self) -> str:
-        if not self._contract:
-            return ""
-        c = self._contract
-        lines = [
-            "Task contract:",
-            f"  archetype: {c.get('task_archetype', 'other')}",
-            f"  language: {c.get('language', 'unknown')}",
-            f"  domain: {c.get('domain', 'general')}",
-            f"  audience: {c.get('audience', 'general')}",
-            f"  output anchor: {c.get('output_format_anchor', 'free-form')}",
-        ]
-        fm = c.get('failure_modes') or []
-        if fm:
-            lines.append(f"  likely failure modes: {'; '.join(str(x) for x in fm[:5])}")
-        rules = c.get('domain_rules') or []
-        if rules:
-            lines.append("  hard domain rules:")
-            for rule in rules[:6]:
-                lines.append(f"    - {rule}")
-        dims = c.get('quality_dimensions') or []
-        if dims:
-            lines.append(f"  evaluation dimensions: {'; '.join(str(x) for x in dims[:8])}")
-        return "\n".join(lines) + "\n\n"
-
-    def _preservation_block(self) -> str:
-        items = self._contract.get('must_preserve') or []
-        original = getattr(self, "_original_prompt", "") or ""
-        concrete: List[str] = []
-        if original:
-            artifacts = self._extract_required_artifacts(original)
-            for kind, values in artifacts.items():
-                if kind == 'code_blocks':
-                    concrete.append(f"preserve {len(values)} fenced code block(s) exactly")
-                elif kind == 'inline_code':
-                    concrete.extend(f"preserve inline code token {v}" for v in values[:8])
-                elif kind == 'placeholders':
-                    concrete.extend(f"preserve placeholder {v}" for v in values[:12])
-                elif kind == 'urls':
-                    concrete.extend(f"preserve URL {v}" for v in values[:8])
-                elif kind == 'xml_tag_names':
-                    concrete.extend(f"preserve XML/tag anchor {v}" for v in values[:12])
-                elif kind == 'json_field_names':
-                    concrete.append(f"preserve JSON field names: {', '.join(values[:20])}")
-                elif kind == 'category_labels':
-                    concrete.append(f"preserve category labels/taxonomy: {', '.join(values[:16])}")
-                elif kind == 'level_values':
-                    concrete.append(f"preserve level scale values: {', '.join(values)}")
-                elif kind == 'enumeration_markers':
-                    concrete.append(f"preserve explicit level/category markers such as: {', '.join(values[:10])}")
-                elif kind == 'enumerated_count':
-                    concrete.append(f"preserve numbered-list structure (original has {values[0]} items)")
-                elif kind == 'markdown_table_lines':
-                    concrete.append("preserve markdown table structure")
-                elif kind == 'geo_escalation_terms':
-                    concrete.append(f"preserve geopolitical escalation terms: {', '.join(values[:10])}")
-        merged_items = list(items)
-        for item in concrete:
-            self._append_unique(merged_items, item)
-        if not merged_items:
-            return ""
-        bullets = "\n".join(f"  - {it}" for it in merged_items[:16])
-        return (
-            "PRESERVATION RULES (must-keep-verbatim artifacts from the original):\n"
-            f"{bullets}\n"
-            "Do NOT rewrite, translate, drop, rename, or paraphrase these artifacts. "
-            "You may improve surrounding prose, but the production contract, schema, "
-            "taxonomy, thresholds, and tags must survive unchanged unless the original "
-            "explicitly asks to change them.\n\n"
-        )
-
-    def _forge_block(self, strategy: str) -> str:
-        top = self._forge.get(strategy) or []
-        if not top:
-            return ""
-        bullets = []
-        for i, snippet in enumerate(top[:2], 1):
-            short = snippet[:180].replace('\n', ' ')
-            if len(snippet) > 180:
-                short += "..."
-            bullets.append(f"  [prev best {i}] {short}")
-        return (
-            f"Your best '{strategy}' outputs from earlier in this run:\n"
-            + "\n".join(bullets)
-            + "\nProduce a new prompt that is AT LEAST as strong as these, ideally stronger.\n\n"
-        )
-
-    def _lessons_block(self) -> str:
-        if not self._lessons:
-            return ""
-        bullets = "\n".join(f"  - {lsn}" for lsn in self._lessons[-4:])
-        return (
-            "Lessons learned so far (why previous winners were strong — reuse these insights):\n"
-            f"{bullets}\n\n"
-        )
-
-    # ══════════════════════════════════════════════════════════════════════
-    # v4.3 — Refusal detection + role-chain fallback
-    # ══════════════════════════════════════════════════════════════════════
-
-    _REFUSAL_RE: Optional[Any] = None
-
-    @classmethod
-    def _is_refusal(cls, text: str) -> bool:
-        """Detect LLM safety refusal (so we can retry on a permissive model)."""
-        if not text or len(text.split()) > 80:
-            return False  # full-length outputs are not refusals
-        if cls._REFUSAL_RE is None:
-            import re as _re
-            cls._REFUSAL_RE = _re.compile("|".join(cls._REFUSAL_PATTERNS), _re.IGNORECASE)
-        return bool(cls._REFUSAL_RE.search(text))
-
-    @staticmethod
-    def _unusable_response_reason(text: str, metadata: Optional[Dict[str, Any]]) -> str:
-        """Return a failure reason if an LLM response should be retried/fallbacked."""
-        meta = metadata or {}
-        error_type = str(meta.get("error_type") or "").lower().strip()
-        if error_type:
-            return error_type
-        finish_reason = str(meta.get("finish_reason") or "").lower().strip()
-        if finish_reason in {"content_filter", "safety", "blocked"}:
-            return "content_filter"
-        if finish_reason in {"length", "max_tokens", "model_length"}:
-            return "length"
-        if finish_reason and finish_reason not in {"stop", "tool_calls", "end_turn"}:
-            return f"finish_reason={finish_reason}"
-        if not (text or "").strip():
-            return "empty"
-        completion_tokens = meta.get("completion_tokens")
-        max_tokens = meta.get("max_tokens")
-        if isinstance(completion_tokens, int) and isinstance(max_tokens, int):
-            if max_tokens > 32 and completion_tokens >= max_tokens - 8:
-                return "near_ceiling"
-        return ""
-
-    def _main_model_chain(self, role: str, model: Optional[str], allow_fallback: bool) -> List[str]:
-        chain = [model] if model else list(self._role_models(role))
-        if allow_fallback and self.FALLBACK_MODEL not in chain:
-            chain.append(self.FALLBACK_MODEL)
-        return self._dedupe_models([m for m in chain if m])
-
-    def _content_filter_fallback_chain(self, chain: List[str]) -> List[str]:
-        if not self._allow_chinese_model_fallbacks():
-            return []
-        return self._dedupe_models(
-            [m for m in self._CONTENT_FILTER_FALLBACK_MODELS if m not in chain],
-            allow_blocked=True,
-        )
-
-    def _generate(
-        self,
-        prompt: str,
-        *,
-        model: Optional[str] = None,
-        role: str = "worker",
-        temperature: float = 0.7,
-        max_tokens: int = 2000,
-        allow_fallback: bool = True,
-    ) -> str:
-        """LLM call through role-specific model chains with finish_reason routing."""
-        max_tokens = max(16, int(max_tokens or 16))
-        main_chain = self._main_model_chain(role, model, allow_fallback)
-        chains = [main_chain]
-        last_resp = ""
-        saw_content_filter = False
-
-        chain_index = 0
-        while chain_index < len(chains):
-            chain = chains[chain_index]
-            for effective_model in chain:
-                model_max_tokens = max_tokens
-                local_attempts = 2 if allow_fallback else 1
-                for local_attempt in range(local_attempts):
-                    try:
-                        gen_kwargs: Dict[str, Any] = {
-                            "prompt": prompt,
-                            "model": effective_model,
-                            "temperature": temperature,
-                            "max_tokens": model_max_tokens,
-                        }
-                        # ULTRA: force max reasoning effort for top-tier
-                        # Anthropic models (Opus 4.7) via OpenRouter extra_body.
-                        if (
-                            self._mode == "ultra"
-                            and effective_model in self._MAX_EFFORT_ULTRA_MODELS
-                        ):
-                            gen_kwargs["extra_body"] = {"reasoning": {"effort": "high"}}
-                        resp = self.llm_client.generate(**gen_kwargs)
-                        metadata = getattr(self.llm_client, "last_response_metadata", {}) or {}
-                        reason = self._unusable_response_reason(resp or "", metadata)
-                    except Exception as exc:
-                        metadata = getattr(self.llm_client, "last_response_metadata", {}) or {}
-                        reason = (
-                            str(metadata.get("error_type") or "")
-                            or getattr(self.llm_client, "last_error_type", None)
-                            or "exception"
-                        )
-                        logger.debug(f"_generate({effective_model}, role={role}) exc: {exc}")
-                        resp = ""
-
-                    last_resp = resp or ""
-                    if last_resp and self._is_refusal(last_resp):
-                        reason = "refusal"
-
-                    if not allow_fallback:
-                        self._record_llm_attempt(
-                            role=role, model=effective_model,
-                            status="success" if not reason else "failed", reason=reason,
-                        )
-                        return last_resp
-
-                    if not reason:
-                        self._record_llm_attempt(role=role, model=effective_model, status="success")
-                        return last_resp
-
-                    self._record_llm_attempt(
-                        role=role, model=effective_model, status="failed", reason=reason,
-                    )
-                    if reason == "content_filter":
-                        saw_content_filter = True
-                    if reason in {"length", "near_ceiling"} and local_attempt == 0:
-                        model_max_tokens = max(model_max_tokens + 256, int(model_max_tokens * 1.35))
-                        continue
-                    if reason == "refusal":
-                        logger.info(
-                            f"_generate: refusal on {effective_model} for role={role}; trying next model"
-                        )
-                    break
-
-            if chain_index == 0 and saw_content_filter:
-                emergency = self._content_filter_fallback_chain(main_chain)
-                if emergency:
-                    chains.append(emergency)
-            chain_index += 1
-
-        return last_resp or ""
-
-    @staticmethod
-    def _schema_to_dict(obj: BaseModel) -> Dict[str, Any]:
-        return obj.model_dump() if hasattr(obj, "model_dump") else dict(obj)
-
-    def _instructor_client(self, model: str) -> Optional[Any]:
-        """Return a cached Instructor client for OpenRouter, or None if unavailable."""
-        try:
-            import instructor  # type: ignore
-        except Exception:
-            return None
-        key = f"openrouter/{model}"
-        if key in self._instructor_clients:
-            return self._instructor_clients[key]
-        kwargs: Dict[str, Any] = {
-            "api_key": self._api_key,
-            "base_url": "https://openrouter.ai/api/v1",
-            "async_client": False,
-        }
-        mode = getattr(getattr(instructor, "Mode", None), "JSON", None)
-        if mode is not None:
-            kwargs["mode"] = mode
-        client = instructor.from_provider(key, **kwargs)
-        self._instructor_clients[key] = client
-        return client
-
-    def _parse_structured_text(
-        self,
-        text: str,
-        schema: Type[BaseModel],
-        *,
-        allowed_starts: Tuple[str, ...],
-    ) -> Optional[BaseModel]:
-        if schema is _JudgeScoreSchema:
-            try:
-                return schema.model_validate({"score": text})
-            except Exception:
-                return None
-        js = self._extract_json_value(text, allowed_starts=allowed_starts)
-        if not js:
-            return None
-        try:
-            value = json.loads(js)
-            if schema is _SyntheticTestsSchema and isinstance(value, list):
-                value = {"tests": value}
-            if schema is _JudgeScoreSchema and isinstance(value, int):
-                value = {"score": value}
-            return schema.model_validate(value)
-        except Exception as exc:
-            logger.debug(f"Structured parse failed for {schema.__name__}: {exc}")
-            return None
-
-    def _generate_structured(
-        self,
-        prompt: str,
-        schema: Type[BaseModel],
-        *,
-        role: str,
-        temperature: float,
-        max_tokens: int,
-        allowed_starts: Tuple[str, ...] = ("{",),
-        max_retries: int = 2,
-    ) -> Optional[BaseModel]:
-        """Instructor-backed structured call with manual JSON/Pydantic fallback."""
-        max_tokens = max(16, int(max_tokens or 16))
-        main_chain = self._main_model_chain(role, None, True)
-        chains = [main_chain]
-        saw_content_filter = False
-
-        chain_index = 0
-        while chain_index < len(chains):
-            chain = chains[chain_index]
-            for effective_model in chain:
-                client = self._instructor_client(effective_model)
-                if client is not None:
-                    try:
-                        obj = client.create(
-                            response_model=schema,
-                            messages=[{"role": "user", "content": prompt}],
-                            temperature=temperature,
-                            max_tokens=max_tokens,
-                            max_retries=max_retries,
-                            extra_body={
-                                "provider": {
-                                    "ignore": ["Google AI Studio"],
-                                    "allow_fallbacks": True,
-                                }
-                            },
-                        )
-                        self._record_llm_attempt(
-                            role=role, model=effective_model, status="success", structured=True,
-                        )
-                        return obj
-                    except Exception as exc:
-                        reason = LLMClient._classify_api_exception(exc)
-                        if reason == "content_filter":
-                            saw_content_filter = True
-                        self._record_llm_attempt(
-                            role=role, model=effective_model,
-                            status="failed", reason=f"instructor:{reason}",
-                            structured=True,
-                        )
-                        logger.debug(
-                            f"Instructor structured call failed on {effective_model} "
-                            f"for {schema.__name__}: {exc}"
-                        )
-
-                # Manual fallback keeps RIDER usable even when Instructor is absent
-                # or a provider/model lacks structured-output/tool support.
-                text = self._generate(
-                    prompt=prompt,
-                    model=effective_model,
-                    role=role,
-                    temperature=temperature,
-                    max_tokens=max_tokens,
-                    allow_fallback=False,
-                )
-                metadata = getattr(self.llm_client, "last_response_metadata", {}) or {}
-                reason = self._unusable_response_reason(text or "", metadata)
-                if reason == "content_filter":
-                    saw_content_filter = True
-                    continue
-                obj = self._parse_structured_text(text or "", schema, allowed_starts=allowed_starts)
-                if obj is not None:
-                    self._record_llm_attempt(
-                        role=role, model=effective_model, status="success", structured=True,
-                    )
-                    return obj
-                self._record_llm_attempt(
-                    role=role, model=effective_model,
-                    status="failed", reason="schema_validation",
-                    structured=True,
-                )
-
-            if chain_index == 0 and saw_content_filter:
-                emergency = self._content_filter_fallback_chain(main_chain)
-                if emergency:
-                    chains.append(emergency)
-            chain_index += 1
-
-        return None
-
-    # ══════════════════════════════════════════════════════════════════════
-    # Primitive helpers
-    # ══════════════════════════════════════════════════════════════════════
-
-    # v4.3 PHASE REACTOR temperatures — 4-phase progression.
-    _PHASE_T = {
-        'ignition': 1.15,       # broad exploration
-        'fusion': 0.85,         # balanced refine
-        'crystallization': 0.55,  # polish, low drift
-        'validation': 0.3,      # stabilize, near-deterministic
-    }
-
-    def _phase_temperature(self, phase: str, default: float = 0.7) -> float:
-        return self._PHASE_T.get(phase, default)
-
-    # v4.4: adaptive complexity detection. Multi-phase pipeline is overkill for
-    # short classification prompts (every refine layer adds noise → bloat),
-    # but essential for long generation prompts (each phase adds structural depth).
-    _SIMPLE_ARCHETYPES = frozenset({
-        'classification', 'data_extraction', 'extractive', 'qa',
-        'yes_no', 'label', 'choice', 'binary',
-    })
-    _COMPLEX_ARCHETYPES = frozenset({
-        'reasoning', 'synthesis', 'generation', 'analysis', 'planning',
-        'writing', 'creative_writing', 'analytical_essay',
-        'technical_explanation', 'code_generation', 'code_review', 'debugging',
-        'summarization', 'translation', 'brainstorming', 'instruction_design',
-        'persuasion', 'creative',
-    })
-
-    def _complexity_tier(self) -> str:
-        """Return 'simple' | 'medium' | 'complex' based on prompt size and archetype.
-
-        v4.4: used by ultra/standard to parametrize pipeline depth. Elite-prompt
-        optimization is not one-size-fits-all — a 2K-char classification prompt
-        needs tight polishing, a 15K-char reasoning prompt needs the full 5-phase
-        beam.
-        """
-        orig_len = getattr(self, '_original_prompt_len', 0) or 0
-        archetype = (self._contract.get('task_archetype') or 'other').lower()
-
-        # Size tier.
-        if orig_len < 3000:
-            size_tier = 'simple'
-        elif orig_len < 8000:
-            size_tier = 'medium'
-        else:
-            size_tier = 'complex'
-
-        # Archetype adjustment: simple archetypes downgrade one level, complex bump up.
-        if archetype in self._SIMPLE_ARCHETYPES:
-            # Large classifiers (>= 7K chars) still need full pipeline — empirically
-            # validated on EXTREM1 (9.7K, classification): full 5-phase gave ideal
-            # hierarchy ultra > blitz > standard > light.
-            if orig_len >= 7000:
-                return 'complex'
-            if size_tier == 'medium':
-                return 'medium'
-            return 'simple'
-        if archetype in self._COMPLEX_ARCHETYPES:
-            if size_tier == 'simple':
-                return 'medium'
-            if size_tier == 'medium':
-                return 'complex'
-            return 'complex'
-        return size_tier
-
-    def _ultra_pipeline_config(self, tier: str) -> Dict[str, Any]:
-        """v4.4: pipeline shape per complexity tier. Fewer phases on simple prompts
-        prevents multi-phase bloat accumulation; all 5 phases fire on complex prompts
-        where each phase provides measurable uplift (verified on EXTREM1 9.7K)."""
-        return {
-            'simple': {
-                # 3 phases: IGNITION + FUSION + SYNTH-EVAL beam. No crystal/validation/red-team.
-                'num_strategies': 3,
-                'tournament_k_phase1': 2,
-                'run_crystallization': False,
-                'run_validation': False,
-                'run_red_team_harden': False,
-                'run_triple_merge': False,
-                'beam_includes_original': True,
-            },
-            'medium': {
-                # 4 phases: IGNITION + FUSION + CRYSTAL + SYNTH-EVAL beam. Skip validation & red-team hardening.
-                'num_strategies': 4,
-                'tournament_k_phase1': 3,
-                'run_crystallization': True,
-                'run_validation': False,
-                'run_red_team_harden': False,
-                'run_triple_merge': False,
-                'beam_includes_original': True,
-            },
-            'complex': {
-                # Full 5 phases + triple-merge. Only fires on long/reasoning/generation prompts.
-                'num_strategies': 5,
-                'tournament_k_phase1': 3,
-                'run_crystallization': True,
-                'run_validation': True,
-                'run_red_team_harden': True,
-                'run_triple_merge': True,
-                'beam_includes_original': True,
-            },
-        }.get(tier, {
-            'num_strategies': 5, 'tournament_k_phase1': 3,
-            'run_crystallization': True, 'run_validation': True,
-            'run_red_team_harden': True, 'run_triple_merge': True,
-            'beam_includes_original': True,
-        })
-
-    # -----------------------------------------------------------------------
-    # Adaptive bloat budget — single source of truth for all 4 bloat-guard
-    # layers (strategy/merge/refine hard caps, synth-eval penalty, beam filter).
-    # -----------------------------------------------------------------------
-    _BLOAT_FLOOR = 4000     # absolute minimum budget — short originals (e.g. 27
-                            # chars) must still be able to expand to a full
-                            # ultra-quality prompt with role/steps/criteria/
-                            # constraints/anti-patterns/examples. 4 KB matches
-                            # the typical length of a well-structured ultra
-                            # prompt (~500-700 words) so short-prompt blow-up
-                            # isn't artificially clipped mid-section.
-    _BLOAT_RATIO = 3.0      # multiplicative cap for long originals — a 10 KB
-                            # prompt should not balloon past 30 KB (runs of
-                            # 10-15x were observed pre-guard on merge pipelines).
-
-    def _bloat_budget(self, orig_len: int) -> int:
-        """Maximum allowed length (in characters) for an optimized prompt.
-
-        Policy:
-          * Short originals (< FLOOR/RATIO = ~833 chars) → FLOOR applies.
-            A vague 27-char prompt is allowed to grow into a 2500-char
-            structured instruction.
-          * Long originals (>= FLOOR/RATIO) → RATIO × original applies.
-            A 4 KB moderation prompt may expand to ~12 KB, a 10 KB
-            rubricator to ~30 KB, but no further.
-
-        Used by _apply_strategy, _merge, _refine (hard cap injected into
-        the worker model's instructions), by _evaluate_candidate_on_tests
-        (synth-eval linear penalty) and by the beam filter before final
-        ranking. Keeping a single helper avoids the 4 layers drifting
-        apart, which was the root cause of the "ultra returns original
-        verbatim" regression on short English prompts.
-        """
-        archetype = str(self._contract.get('task_archetype') or '').lower()
-        if (
-            self._is_underoptimized_prompt(getattr(self, "_original_prompt", ""))
-            and archetype in {'creative_writing', 'analytical_essay', 'persuasion', 'brainstorming', 'other'}
-        ):
-            # Short under-specified prompts need room to become real production
-            # instructions. Keep long operational prompts strict; loosen only
-            # for raw under-specified creative/general asks.
-            return max(int(orig_len * 8.0), 6500)
-        return max(int(orig_len * self._BLOAT_RATIO), self._BLOAT_FLOOR)
-
-    def _apply_strategy(
-        self, prompt: str, strategy: str,
-        extra_vars: Optional[Dict[str, str]] = None,
-        temperature: float = 0.7,
-    ) -> Optional[str]:
-        """Apply one improvement strategy. 1 call. Returns improved prompt or None on failure."""
-        template = self._STRATEGY_PROMPTS.get(strategy)
-        if template is None:
-            return None
-        vars_ = {
-            'prompt': prompt,
-            'contract_block': self._contract_block(),
-            'preservation_block': self._preservation_block(),
-            'forge_block': self._forge_block(strategy),
-            'lessons_block': self._lessons_block(),
-            'collapsed_preview': '',
-        }
-        if extra_vars:
-            vars_.update(extra_vars)
-        meta = template.format(**vars_)
-        # Hard char budget so strategies don't balloon.
-        orig_ref = getattr(self, "_original_prompt_len", 0) or len(prompt)
-        char_budget = self._bloat_budget(orig_ref)
-        meta += f"\n\nHARD LIMIT: the improved prompt must not exceed {char_budget} characters."
-        try:
-            resp = self._generate(
-                prompt=meta, role=self._strategy_role(strategy), temperature=temperature,
-                max_tokens=self._adaptive_max_tokens(prompt),
-            )
-            text = self._strip_wrappers(resp)
-            if len(text.split()) < 15:
-                return None
-            # Update FORGE memory with this success.
-            self._forge.setdefault(strategy, []).append(text)
-            # Keep only top-3 by length proxy (longer usually = more detail); dedup.
-            self._forge[strategy] = list(dict.fromkeys(self._forge[strategy]))[-3:]
-            return text
-        except Exception as exc:
-            logger.debug(f"_apply_strategy({strategy}) failed: {exc}")
-            return None
-
-    @staticmethod
-    def _strip_wrappers(text: str) -> str:
-        if text is None:
-            return ""
-        text = text.strip()
-        for tag in ('<prompt>', '</prompt>', '[PROMPT_START]', '[PROMPT_END]',
-                    '<START>', '<END>', '```text', '```markdown', '```'):
-            text = text.replace(tag, '')
-        return text.strip().strip('"').strip("'").strip()
-
-    def _select_best(
-        self, candidates: List[Tuple[str, str]], collect_lesson: bool = True,
-    ) -> Tuple[str, str]:
-        """Pairwise select best + extract GENESIS-lite lesson. 1 Sonnet call."""
-        if len(candidates) <= 1:
-            return candidates[0]
-        parts = []
-        for i, (_, text) in enumerate(candidates, 1):
-            parts.append(f"--- PROMPT {i} ---\n{text}")
-        meta = self._COMPARE_PROMPT.format(
-            contract_block=self._contract_block(),
-            candidates="\n\n".join(parts),
-        )
-        try:
-            resp = self._generate(
-                prompt=meta, role="judge",
-                temperature=0.0, max_tokens=180,
-                allow_fallback=True,
-            )
-            winner_idx, why = self._parse_winner_why(resp, len(candidates))
-            if winner_idx is not None:
-                if collect_lesson and why:
-                    self._lessons.append(why.strip())
-                return candidates[winner_idx]
-        except Exception as exc:
-            logger.debug(f"_select_best failed: {exc}")
-        for n, t in candidates:
-            if n != "original":
-                return (n, t)
-        return candidates[0]
-
-    def _rank_batch(
-        self, candidates: List[Tuple[str, str]], k: int = 3,
-        collect_lesson: bool = True,
-    ) -> List[Tuple[str, str]]:
-        """v4.3: rank N candidates in ONE Sonnet call (N-way tournament).
-
-        Replaces 3 sequential _select_best calls when selecting top-k from N.
-        Returns top-k (name, text) tuples in order.
-        """
-        if len(candidates) <= k:
-            return list(candidates)
-        parts = []
-        for i, (_, text) in enumerate(candidates, 1):
-            parts.append(f"--- PROMPT {i} ---\n{text}")
-        meta = self._BATCH_RANK_PROMPT.format(
-            contract_block=self._contract_block(),
-            n=len(candidates),
-            candidates="\n\n".join(parts),
-        )
-        try:
-            resp = self._generate(
-                prompt=meta, role="judge",
-                temperature=0.0, max_tokens=240,
-                allow_fallback=True,
-            )
-            ranked_idx, why_top = self._parse_ranked(resp, len(candidates))
-            if ranked_idx and len(ranked_idx) >= k:
-                if collect_lesson and why_top:
-                    self._lessons.append(why_top.strip())
-                return [candidates[i] for i in ranked_idx[:k]]
-        except Exception as exc:
-            logger.debug(f"_rank_batch failed: {exc}")
-        # Fallback: sequential _select_best
-        top: List[Tuple[str, str]] = []
-        remaining = list(candidates)
-        for _ in range(min(k, len(remaining))):
-            w = self._select_best(remaining, collect_lesson=collect_lesson and not top)
-            top.append(w)
-            remaining = [(n, t) for n, t in remaining if t != w[1]]
-        return top
-
-    @staticmethod
-    def _parse_ranked(resp: str, n_cands: int) -> Tuple[List[int], Optional[str]]:
-        """Parse 'RANKED: 3,1,4,2\\nWHY_TOP: ...' response into 0-indexed idx list + why."""
-        if not resp:
-            return [], None
-        m_r = re.search(r'RANKED\s*:\s*([\d,\s]+)', resp, re.IGNORECASE)
-        idx_list: List[int] = []
-        seen: set = set()
-        if m_r:
-            for part in re.findall(r'\d+', m_r.group(1)):
-                idx = int(part) - 1
-                if 0 <= idx < n_cands and idx not in seen:
-                    idx_list.append(idx)
-                    seen.add(idx)
-        # If RANKED parse failed, try to pick all digits from resp.
-        if not idx_list:
-            for part in re.findall(r'\b(\d+)\b', resp):
-                idx = int(part) - 1
-                if 0 <= idx < n_cands and idx not in seen:
-                    idx_list.append(idx)
-                    seen.add(idx)
-        m_y = re.search(r'WHY_TOP\s*:\s*(.+?)(?:\n|$)', resp, re.IGNORECASE | re.DOTALL)
-        why = None
-        if m_y:
-            why = re.sub(r'\s+', ' ', m_y.group(1)).strip()
-            if len(why) > 240:
-                why = why[:240].rstrip() + "..."
-        return idx_list, why
-
-    @staticmethod
-    def _parse_winner_why(resp: str, n_cands: int) -> Tuple[Optional[int], Optional[str]]:
-        """Parse 'WINNER: X\nWHY: ...' response."""
-        if not resp:
-            return None, None
-        winner_idx: Optional[int] = None
-        why: Optional[str] = None
-        m_w = re.search(r'WINNER\s*:\s*(\d+)', resp, re.IGNORECASE)
-        if m_w:
-            idx = int(m_w.group(1)) - 1
-            if 0 <= idx < n_cands:
-                winner_idx = idx
-        if winner_idx is None:
-            # Fallback: first bare digit in [1..N].
-            for num in re.findall(r'\b(\d+)\b', resp):
-                idx = int(num) - 1
-                if 0 <= idx < n_cands:
-                    winner_idx = idx
-                    break
-        m_y = re.search(r'WHY\s*:\s*(.+?)(?:\n|$)', resp, re.IGNORECASE | re.DOTALL)
-        if m_y:
-            why = re.sub(r'\s+', ' ', m_y.group(1)).strip()
-            if len(why) > 240:
-                why = why[:240].rstrip() + "..."
-        return winner_idx, why
-
-    def _merge(self, p1: str, p2: str, temperature: float = 0.5) -> str:
-        """Merge two prompts. 1 call.
-
-        Hard char budget via _bloat_budget() — short originals get absolute floor
-        (can expand to ultra-quality), long originals capped at 3x (no runaway).
-        """
-        orig_ref = getattr(self, "_original_prompt_len", 0) or max(len(p1), len(p2))
-        char_budget = self._bloat_budget(orig_ref)
-        meta = self._MERGE_PROMPT.format(
-            prompt_a=p1, prompt_b=p2,
-            contract_block=self._contract_block(),
-            preservation_block=self._preservation_block(),
-            lessons_block=self._lessons_block(),
-        )
-        meta += f"\n\nHARD LIMIT: the merged prompt must not exceed {char_budget} characters."
-        try:
-            resp = self._generate(
-                prompt=meta, role="worker", temperature=temperature,
-                max_tokens=self._adaptive_max_tokens(p1, p2),
-            )
-            text = self._strip_wrappers(resp)
-            return text if len(text.split()) >= 15 else p1
-        except Exception:
-            return p1
-
-    def _constitutional_audit(self, prompt: str) -> str:
-        """6-dim rubric audit. 1 Sonnet call. Returns TOP_WEAKNESSES block as directive text."""
-        meta = self._CONSTITUTIONAL_AUDIT_PROMPT.format(
-            prompt=prompt, contract_block=self._contract_block(),
-        )
-        try:
-            resp = self._generate(
-                prompt=meta, role="critic",
-                temperature=0.2, max_tokens=700,
-            )
-            # Extract TOP_WEAKNESSES section if present; else return full response.
-            m = re.search(r'TOP_WEAKNESSES\s*:\s*(.+)$', resp, re.IGNORECASE | re.DOTALL)
-            directives = m.group(1).strip() if m else resp.strip()
-            return directives or "Prompt needs more specificity and tighter constraints."
-        except Exception:
-            return "Prompt needs more specificity, clearer role, and explicit output format."
-
-    def _refine(self, prompt: str, weaknesses: str, temperature: float = 0.5) -> str:
-        """Refine based on audit directives. 1 call.
-
-        Hard char budget via _bloat_budget() — short originals get absolute floor
-        (can expand to ultra-quality), long originals capped at 3x (no runaway).
-        """
-        orig_ref = getattr(self, "_original_prompt_len", 0) or len(prompt)
-        char_budget = self._bloat_budget(orig_ref)
-        meta = self._REFINE_PROMPT.format(
-            prompt=prompt, weaknesses=weaknesses,
-            contract_block=self._contract_block(),
-            preservation_block=self._preservation_block(),
-            lessons_block=self._lessons_block(),
-        )
-        meta += f"\n\nHARD LIMIT: the refined prompt must not exceed {char_budget} characters."
-        try:
-            resp = self._generate(
-                prompt=meta, role="worker", temperature=temperature,
-                max_tokens=self._adaptive_max_tokens(prompt),
-            )
-            text = self._strip_wrappers(resp)
-            return text if len(text.split()) >= 15 else prompt
-        except Exception:
-            return prompt
-
-    def _estimate_quality(self, original: str, improved: str) -> Tuple[float, float]:
-        """Rate original vs improved. 1 Sonnet call. Returns (orig, improved) scores in [0,1].
-
-        v4.3: if a real synthetic-eval winner score was measured during this run
-        (self._synth_best_score), use it directly as the improved-side fitness.
-        """
-        # v4.3: prefer real synthetic-eval score over rubric heuristic.
-        synth = getattr(self, "_synth_best_score", None)
-        synth_orig = getattr(self, "_synth_orig_score", None)
-        if synth is not None and synth > 0 and synth_orig is not None and synth_orig > 0:
-            return float(synth_orig), float(synth)
-        meta = self._QUALITY_PROMPT.format(
-            prompt_a=original, prompt_b=improved,
-            contract_block=self._contract_block(),
-        )
-        try:
-            resp = self._generate(
-                prompt=meta, role="judge",
-                temperature=0.0, max_tokens=30,
-            )
-            nums = re.findall(r'[AB]\s*:\s*(\d+)', resp)
-            if len(nums) >= 2:
-                a = min(10, max(1, int(nums[0]))) / 10.0
-                b = min(10, max(1, int(nums[1]))) / 10.0
-                return a, b
-        except Exception:
-            pass
-        return 0.3, 0.7
-
-    # ══════════════════════════════════════════════════════════════════════
-    # Preservation validator (post-hoc, no LLM; repair is 1 conditional call)
-    # ══════════════════════════════════════════════════════════════════════
-
-    _CODE_FENCE_RE = re.compile(r"```[\s\S]*?```")
-    _INLINE_CODE_RE = re.compile(r"`[^`\n]+`")
-    _PLACEHOLDER_RE = re.compile(r"\{\{[^{}\s][^{}]*\}\}|\{[A-Za-z_][\w\.]*\}")
-    _URL_RE = re.compile(r"https?://[^\s)>\]]+", re.IGNORECASE)
-    _MD_TABLE_RE = re.compile(r"^\s*\|.+\|\s*$", re.MULTILINE)
-    _JSON_FIELD_RE = re.compile(r'"([A-Za-z_][A-Za-z0-9_]*)"\s*:')
-    _LEGAL_CITATION_RE = re.compile(
-        r"(?:ст\.?\s*\d+(?:\.\d+)?|ФЗ-?\s*№?\s*\d+|№\s*\d+-ФЗ|КоАП\s+РФ|УК\s+РФ)",
-        re.IGNORECASE,
-    )
-    _GENERIC_PREFIXES = (
-        r"write\s+(?:a|an|the)?\s*",
-        r"create\s+(?:a|an|the)?\s*",
-        r"make\s+(?:a|an|the)?\s*",
-        r"help\s+me\s+",
-        r"напиши\s+",
-        r"сделай\s+",
-        r"создай\s+",
-        r"помоги\s+",
-        r"напиши\s+",
-        r"сделай\s+",
-        r"создай\s+",
-        r"помоги\s+",
-    )
-
-    @classmethod
-    def _is_underoptimized_prompt(cls, text: str) -> bool:
-        """Detect short/raw prompts where returning original is never impressive."""
-        stripped = (text or '').strip()
-        if not stripped:
-            return False
-        words = re.findall(r"\w+", stripped, re.UNICODE)
-        if len(stripped) <= 180:
-            return True
-        if len(words) <= 35 and not re.search(r"\n|```|<[^>]+>|{[A-Za-z_]", stripped):
-            return True
-        lower = stripped.lower()
-        return any(re.match(pat, lower) for pat in cls._GENERIC_PREFIXES)
-    # Custom XML-like tags used as structural anchors (e.g. <шкала_нарушений>). Allows
-    # Cyrillic tag names used in RU prompt engineering (moderation rubrics).
-    # Matches opening/closing tag pair: <name>...</name>, name length ≥ 3 to avoid
-    # catching generic HTML stubs like <p>.
-    _XML_TAG_RE = re.compile(r"<([\w\u0400-\u04FF_]{3,})>[\s\S]*?</\1>")
-
-    def _adaptive_max_tokens(self, *prompts: str, min_tokens: int = 4000, ceiling: int = 16000) -> int:
-        """v4.3: Cyrillic-aware scaling so full rewrites don't get truncated, but
-        also don't balloon 10x — cap at input*1.5 + 1024 headroom.
-
-        Cyrillic text tokenizes ~2x worse than English in Anthropic/OpenAI tokenizers.
-        Previous v4.2 defaults (min=2000) caused mid-sentence truncation on 4K+ char
-        Russian prompts. v4.3.1 raised ceiling to 24K which let ultra balloon to 24K
-        chars (7x original). v4.3.2: tighter cap — enough room for 50% growth + structure.
-        """
-        total_chars = 0
-        cyr_chars = 0
-        for p in prompts:
-            if not p:
-                continue
-            total_chars += len(p)
-            for ch in p:
-                code = ord(ch.lower())
-                if 0x0400 <= code <= 0x04FF:
-                    cyr_chars += 1
-        cyr_ratio = (cyr_chars / total_chars) if total_chars else 0.0
-        if cyr_ratio > 0.25:
-            tokens_per_char = 2.0
-        elif cyr_ratio > 0.1:
-            tokens_per_char = 1.2
-        else:
-            tokens_per_char = 0.3
-        est_input_tokens = int(total_chars * tokens_per_char)
-        # Output: input x1.2 (+20% growth allowance for structural additions) + 1024 headroom.
-        budget = int(est_input_tokens * 1.2) + 1024
-        return max(min_tokens, min(ceiling, budget))
-
-    def _extract_required_artifacts(self, original: str) -> Dict[str, List[str]]:
-        """Detect artifacts in the original that should be preserved."""
-        artifacts: Dict[str, List[str]] = {}
-        preserve_items = [str(x) for x in (self._contract.get('must_preserve') or [])]
-        flags = {x.lower() for x in preserve_items}
-        # Auto-detect common artifacts regardless of contract flags.
-        code_fences = self._CODE_FENCE_RE.findall(original)
-        if code_fences:
-            artifacts['code_blocks'] = code_fences
-        inline_code = self._INLINE_CODE_RE.findall(original)
-        if inline_code:
-            artifacts['inline_code'] = inline_code
-        placeholders = self._PLACEHOLDER_RE.findall(original)
-        if placeholders:
-            artifacts['placeholders'] = placeholders
-        urls = self._URL_RE.findall(original)
-        if urls:
-            artifacts['urls'] = urls
-        # Markdown tables — any markdown/table mention in contract or auto-detect.
-        if any('table' in f or 'markdown' in f for f in flags) or self._MD_TABLE_RE.search(original):
-            lines = [ln for ln in original.splitlines() if self._MD_TABLE_RE.match(ln)]
-            if lines:
-                artifacts['markdown_table_lines'] = lines
-        # Custom XML-like tags (e.g. <шкала_нарушений>, <scale>, etc.) — common anchor
-        # in moderation / rubric prompts.
-        xml_matches = [m.group(0) for m in self._XML_TAG_RE.finditer(original)]
-        if xml_matches:
-            artifacts['xml_anchors'] = xml_matches
-            # Also require the opening and closing tag names to reappear.
-            tag_names: List[str] = []
-            for m in self._XML_TAG_RE.finditer(original):
-                name = m.group(1)
-                if name not in tag_names:
-                    tag_names.append(f"<{name}>")
-                    tag_names.append(f"</{name}>")
-            if tag_names:
-                artifacts['xml_tag_names'] = tag_names
-        json_fields: List[str] = []
-        for field in self._JSON_FIELD_RE.findall(original):
-            if field not in json_fields:
-                json_fields.append(field)
-        if json_fields:
-            artifacts['json_field_names'] = json_fields[:80]
-        category_labels: List[str] = []
-        for pat in (
-            r'"name"\s*:\s*"([^"]+)"',
-            r'"filter"\s*:\s*"([^"]+)"',
-            r"\b(?:name|filter)\s*=\s*['\"]([^'\"]+)['\"]",
-            r"\b[A-ZА-ЯЁ0-9]+_[A-ZА-ЯЁ0-9_]+\b",
-        ):
-            for m in re.finditer(pat, original):
-                value = m.group(1) if m.groups() else m.group(0)
-                if 2 < len(value) <= 80 and value not in category_labels:
-                    category_labels.append(value)
-        if category_labels:
-            artifacts['category_labels'] = category_labels[:80]
-        level_values = []
-        for m in re.finditer(r'\b(?:level|уровень)[ \t]*(?:=|:|-)?[ \t]*([0-3])\b', original, re.IGNORECASE):
-            value = m.group(1)
-            if value not in level_values:
-                level_values.append(value)
-        if level_values:
-            artifacts['level_values'] = level_values
-        # Director-feedback invariants from real EKSMO moderation prompts:
-        # the optimizer may improve wording, but it must not soften the machine
-        # output contract, priority order, level caps, or exception semantics.
-        policy_lines = [ln.strip() for ln in original.splitlines() if 4 <= len(ln.strip()) <= 500]
-        exclusive_output_lines: List[str] = []
-        level_policy_lines: List[str] = []
-        priority_policy_lines: List[str] = []
-        for line in policy_lines:
-            low = line.lower()
-            if any(term in low for term in (
-                'единственный допустимый вывод',
-                'единственный вывод',
-                'только json',
-                'строго json',
-                'без пояснений',
-                'без объяснений',
-                'only allowed output',
-                'output only',
-                'json only',
-                'no prose',
-                'no explanations',
-            )):
-                if line not in exclusive_output_lines:
-                    exclusive_output_lines.append(line)
-            has_level_word = ('level' in low) or ('уров' in low)
-            if has_level_word and any(term in low for term in (
-                'не выше',
-                'выше',
-                'поднима',
-                'повыш',
-                'кроме',
-                'исключ',
-                'нулев',
-                'level 0',
-                'level 3',
-                'cap',
-                'maximum',
-                'max ',
-                'raise',
-                'escalat',
-                'except',
-            )):
-                if line not in level_policy_lines:
-                    level_policy_lines.append(line)
-            if any(term in low for term in ('приоритет', 'порядок', 'сначала', 'затем', 'priority', 'order')) and (
-                has_level_word or 'категор' in low or 'category' in low or 'rule' in low
-            ):
-                if line not in priority_policy_lines:
-                    priority_policy_lines.append(line)
-        if exclusive_output_lines:
-            artifacts['exclusive_output_rules'] = exclusive_output_lines[:12]
-        if level_policy_lines:
-            artifacts['level_escalation_rules'] = level_policy_lines[:20]
-        if priority_policy_lines:
-            artifacts['priority_policy_rules'] = priority_policy_lines[:12]
-        legal_citations: List[str] = []
-        for m in self._LEGAL_CITATION_RE.finditer(original):
-            value = m.group(0)
-            if value not in legal_citations:
-                legal_citations.append(value)
-        if legal_citations:
-            artifacts['legal_citations'] = legal_citations[:40]
-        geo_terms: List[str] = []
-        for term in ('Россия', 'Украина', 'РФ', 'СВО', 'ВС РФ', 'Вооруженных сил', 'Вооружённых сил'):
-            if term.lower() in original.lower():
-                geo_terms.append(term)
-        if geo_terms:
-            artifacts['geo_escalation_terms'] = geo_terms
-        # Literal strings from contract.must_preserve that look like concrete quotes /
-        # markers (not generic categories). A simple heuristic: if the item appears
-        # verbatim in the original, treat it as a hard preservation requirement.
-        literal_hits: List[str] = []
-        for item in preserve_items:
-            item_clean = item.strip()
-            if len(item_clean) < 3 or len(item_clean) > 200:
-                continue
-            if item_clean in original and item_clean not in literal_hits:
-                literal_hits.append(item_clean)
-        if literal_hits:
-            artifacts['contract_literal_strings'] = literal_hits
-        # v4.1 Enumerated structure preservation — the #1 failure mode for legal /
-        # moderation rubrics. LLM judges prefer prose flow, but lawyers need
-        # numbered criteria for citation in court.
-        numbered_items = re.findall(r"(?m)^\s*\d+[\.\)]\s+\S", original)
-        if len(numbered_items) >= 3:
-            artifacts['enumerated_count'] = [str(len(numbered_items))]
-        # Explicit level / step markers (case-insensitive, multi-language).
-        # `[ \t]+` (no newlines) to avoid matching word-across-line artifacts.
-        marker_patterns = [
-            r"\b(?:Уровень|Ступень|Level|Step|ШАГ|Этап|Stage|Tier)[ \t]+\d+(?:\.\d+)?",
-            # Category codes like "1.1.а", "2.3.б" (cyrillic or latin letter suffix).
-            r"\b\d+\.\d+(?:\.[a-zа-я])?",
-            # v4.3: Letter-based markers used in legal/moderation rubrics.
-            # Matches "ПРИЗНАК A", "MARKER B", "ITEM C", "CRITERION D", "ПУНКТ А" etc.
-            r"\b(?:ПРИЗНАК|Признак|MARKER|Marker|ITEM|Item|CRITERION|Criterion|ПУНКТ|Пункт|КРИТЕРИЙ|Критерий)[ \t]+[A-ZА-Я]\b",
-        ]
-        markers: List[str] = []
-        for pat in marker_patterns:
-            for m in re.finditer(pat, original, re.IGNORECASE):
-                txt = m.group(0)
-                if txt not in markers:
-                    markers.append(txt)
-        if markers:
-            artifacts['enumeration_markers'] = markers[:50]  # cap for safety
-        return artifacts
-
-    def _check_preservation(self, original: str, improved: str) -> Tuple[List[str], Dict[str, List[str]]]:
-        """Return (violations, required_items_by_kind)."""
-        artifacts = self._extract_required_artifacts(original)
-        violations: List[str] = []
-        for kind, items in artifacts.items():
-            if kind == 'enumerated_count':
-                # Special: numbered list count must not drop by more than 40%.
-                try:
-                    expected_min = int(int(items[0]) * 0.6)
-                except Exception:
-                    continue
-                actual = len(re.findall(r"(?m)^\s*\d+[\.\)]\s+\S", improved))
-                if actual < expected_min:
-                    violations.append(
-                        f"enumerated_structure_loss: expected at least {expected_min} "
-                        f"numbered items, found {actual} (original had {items[0]})"
-                    )
-                continue
-            if kind == 'enumeration_markers':
-                # All explicit level/step/category markers must remain present.
-                missing = [m for m in items if m not in improved]
-                if missing:
-                    for mk in missing[:6]:
-                        violations.append(f"missing enumeration_marker: {mk}")
-                continue
-            if kind == 'json_field_names':
-                improved_fields = set(self._JSON_FIELD_RE.findall(improved))
-                missing = [m for m in items if m not in improved_fields]
-                if missing:
-                    violations.append(f"missing json_field_names: {', '.join(missing[:12])}")
-                continue
-            if kind == 'category_labels':
-                missing = [m for m in items if m not in improved]
-                if missing:
-                    violations.append(f"missing category_labels: {', '.join(missing[:8])}")
-                continue
-            if kind == 'level_values':
-                improved_values = set(
-                    m.group(1) for m in re.finditer(
-                        r'\b(?:level|уровень)[ \t]*(?:=|:|-)?[ \t]*([0-3])\b',
-                        improved,
-                        re.IGNORECASE,
-                    )
-                )
-                missing = [m for m in items if m not in improved_values]
-                if missing:
-                    violations.append(f"missing level_values: {', '.join(missing)}")
-                continue
-            if kind == 'legal_citations':
-                missing = [m for m in items if m not in improved]
-                if missing:
-                    violations.append(f"missing legal_citations: {', '.join(missing[:6])}")
-                continue
-            if kind == 'geo_escalation_terms':
-                missing = [m for m in items if m.lower() not in improved.lower()]
-                if missing:
-                    violations.append(f"missing geo_escalation_terms: {', '.join(missing[:6])}")
-                continue
-            for item in items:
-                if item and item not in improved:
-                    violations.append(f"missing {kind}: {item[:120]}")
-        return violations, artifacts
-
-    def _repair_preservation(
-        self, original: str, improved: str, violations: List[str],
-        required: Dict[str, List[str]],
-    ) -> str:
-        """One repair call — only fired when violations were detected."""
-        if not violations:
-            return improved
-        flat_items: List[str] = []
-        for kind, items in required.items():
-            for it in items[:4]:
-                flat_items.append(f"[{kind}] {it}")
-        required_text = "\n".join(flat_items[:12]) or "(auto-detected artifacts above)"
-        meta = self._PRESERVE_REPAIR_PROMPT.format(
-            required_items=required_text,
-            original=original, broken=improved,
-            violations="\n".join(f"- {v}" for v in violations[:12]),
-        )
-        try:
-            resp = self._generate(
-                prompt=meta, role="worker", temperature=0.1,
-                max_tokens=self._adaptive_max_tokens(original, improved),
-            )
-            text = self._strip_wrappers(resp)
-            return text if len(text.split()) >= 15 else improved
-        except Exception:
-            return improved
-
-    @staticmethod
-    def _check_completeness_issues(text: str) -> List[str]:
-        """Conservative truncation/completeness checks for generated prompts."""
-        issues: List[str] = []
-        if not text or not text.strip():
-            return ["empty_prompt"]
-        stripped = text.rstrip()
-        if stripped.count("```") % 2:
-            issues.append("unclosed_code_fence")
-
-        tag_re = re.compile(r"</?([\w\u0400-\u04FF_]{3,})(?:\s[^>]*)?>")
-        opens: Dict[str, int] = {}
-        closes: Dict[str, int] = {}
-        structural_opens: Dict[str, int] = {}
-        structural_tag_names = {'перевод', 'translation', 'rules', 'final_check', 'шкала_нарушений'}
-        for m in tag_re.finditer(stripped):
-            raw = m.group(0)
-            if raw.endswith("/>"):
-                continue
-            name = m.group(1)
-            if raw.startswith("</"):
-                closes[name] = closes.get(name, 0) + 1
-            else:
-                opens[name] = opens.get(name, 0) + 1
-                line_start = stripped.rfind("\n", 0, m.start()) + 1
-                line_end = stripped.find("\n", m.end())
-                if line_end == -1:
-                    line_end = len(stripped)
-                line = stripped[line_start:line_end].strip()
-                if line == raw or name in structural_tag_names:
-                    structural_opens[name] = structural_opens.get(name, 0) + 1
-        for name in opens:
-            structural_count = structural_opens.get(name, 0)
-            if (name in structural_tag_names or closes.get(name, 0) > 0) and structural_count and closes.get(name, 0) < structural_count:
-                issues.append(f"unclosed_xml_tag:{name}")
-                if len(issues) >= 4:
-                    break
-
-        last_line = next((ln.strip() for ln in reversed(stripped.splitlines()) if ln.strip()), "")
-        last_word_match = re.search(r"([\w\u0400-\u04FF-]+)$", last_line)
-        last_word = (last_word_match.group(1).lower() if last_word_match else "")
-        dangling_words = {
-            'if', 'when', 'while', 'because', 'and', 'or', 'with', 'for',
-            'если', 'когда', 'пока', 'потому', 'что', 'как', 'и', 'или',
-            'эпизод', 'упоминание', 'критер', 'критерий', 'раздел', 'пункт',
-            'если', 'когда', 'пока', 'потому', 'что', 'как', 'и', 'или',
-            'эпизод', 'упоминание', 'критер', 'критерий', 'раздел', 'пункт',
-        }
-        if re.match(r"^\s*#{1,6}\s+\S", last_line) and not re.search(r"[.!?:;…)]$", last_line):
-            issues.append("ends_with_heading")
-        if last_word in dangling_words or re.search(r"[,;({\[-]\s*$", last_line):
-            issues.append("dangling_final_line")
-        return issues
-
-    def _validate_and_repair(self, original: str, improved: str) -> str:
-        violations, required = self._check_preservation(original, improved)
-        if not violations:
-            return improved
-        self._log(f"  [preservation] {len(violations)} violation(s), repairing")
-        return self._repair_preservation(original, improved, violations, required)
-
-    def _deterministic_safe_enhancement(self, original: str) -> str:
-        """Non-LLM fallback that improves the wrapper while preserving source verbatim."""
-        archetype = str(self._contract.get('task_archetype') or '').lower()
-        lang = str(self._contract.get('language') or '').lower()
-        is_ru = lang == 'ru' or bool(re.search(r'[\u0400-\u04FF]', original or ''))
-
-        if archetype in {'debugging', 'code_generation', 'code_review'}:
-            template = RIDER_SAFE_CODE_PROMPT_RU if is_ru else RIDER_SAFE_CODE_PROMPT_EN
-            return template.format(original=original)
-
-        if archetype == 'classification':
-            template = (
-                RIDER_SAFE_CLASSIFICATION_PROMPT_RU
-                if is_ru else
-                RIDER_SAFE_CLASSIFICATION_PROMPT_EN
-            )
-            return template.format(original=original)
-
-        if archetype == 'translation':
-            template = RIDER_SAFE_TRANSLATION_PROMPT_RU if is_ru else RIDER_SAFE_TRANSLATION_PROMPT_EN
-            return template.format(original=original)
-
-        if is_ru:
-            return RIDER_SAFE_GENERIC_PROMPT_RU.format(original=original)
-        return RIDER_SAFE_GENERIC_PROMPT_EN.format(original=original)
-
-    # ══════════════════════════════════════════════════════════════════════
-    # Diversity helpers (Jaccard, no embeddings)
-    # ══════════════════════════════════════════════════════════════════════
-
-    _WORD_RE = re.compile(r"\w+", re.UNICODE)
-    _STOPWORDS_EN = {
-        'the', 'a', 'an', 'and', 'or', 'of', 'to', 'in', 'for', 'on', 'with', 'as',
-        'by', 'is', 'are', 'be', 'been', 'was', 'were', 'this', 'that', 'it', 'at',
-        'from', 'you', 'your', 'we', 'our', 'they', 'them', 'not', 'no', 'do', 'does',
-        'should', 'must', 'can', 'will', 'would', 'could',
-    }
-
-    @classmethod
-    def _tokens(cls, text: str) -> set:
-        words = [w.lower() for w in cls._WORD_RE.findall(text or '')]
-        return {w for w in words if len(w) > 2 and w not in cls._STOPWORDS_EN}
-
-    @classmethod
-    def _jaccard(cls, a: str, b: str) -> float:
-        ta, tb = cls._tokens(a), cls._tokens(b)
-        if not ta or not tb:
-            return 0.0
-        inter = len(ta & tb)
-        union = len(ta | tb)
-        return inter / union if union else 0.0
-
-    def _pick_diverse_pair(
-        self, candidates: List[Tuple[str, str]], similarity_threshold: float = 0.85,
-    ) -> Tuple[Optional[Tuple[str, str]], Optional[Tuple[str, str]], bool]:
-        """Pick top-1 + most diverse partner. Returns (top1, partner, is_collapsed)."""
-        if not candidates:
-            return None, None, True
-        top1 = self._select_best(candidates, collect_lesson=True)
-        rest = [c for c in candidates if c[1] != top1[1]]
-        if not rest:
-            return top1, None, True
-        # Choose the candidate with LOWEST Jaccard vs top1 AND still high enough quality
-        # (we already excluded top1 — rest contains the rest of the arena).
-        scored = [(c, self._jaccard(top1[1], c[1])) for c in rest]
-        scored.sort(key=lambda x: x[1])  # ascending similarity
-        partner, min_sim = scored[0]
-        collapsed = min_sim >= similarity_threshold
-        return top1, partner, collapsed
-
-    # ══════════════════════════════════════════════════════════════════════
-    # Run plumbing
-    # ══════════════════════════════════════════════════════════════════════
-
-    def _setup_run(self, prompt: str, mode: str):
-        self._api_calls_start = self.llm_client.total_api_calls
-        self._history = []
-        self._final_prompt = prompt
-        self._original_fitness = 0.0
-        self._best_fitness = 0.0
-        self._contract = {}
-        self._lessons = []
-        self._forge = {}
-        self._mode = mode
-        self._original_prompt = prompt or ""
-        self._role_model_chains = self._build_role_model_chains(mode)
-        self.model = self._role_model("worker")
-        self._synthetic_tests = []
-        self._synthetic_rankings = []
-        self._llm_attempts = []
-        # v4.3 real-quality tracking
-        self._synth_best_score: Optional[float] = None
-        self._synth_orig_score: Optional[float] = None
-        self._phase_temps_used: List[str] = []
-        # v4.3.2 bloat guard reference length
-        self._original_prompt_len: int = len(prompt or "")
-        self._log("=" * 60)
-        self._log(f"RIDER Genesis [{mode.upper()}]")
-        self._log("=" * 60)
-        self._log("Models: " + " | ".join(
-            f"{role}={self._role_model(role)}" for role in self._ROLES
-        ))
-        self._log(f"Original: {prompt[:100]}{'...' if len(prompt) > 100 else ''}")
-
-    def _finalize_run(self, prompt: str, result: str, mode: str, t0: float) -> str:
-        # Preservation check — may add at most 1 call.
-        result = self._validate_and_repair(prompt, result)
-        final_violations, _ = self._check_preservation(prompt, result)
-        completeness_issues = self._check_completeness_issues(result)
-        if final_violations or completeness_issues:
-            if completeness_issues:
-                self._log(f"  [safety] completeness issues: {completeness_issues[:4]}")
-            fallback = self._deterministic_safe_enhancement(prompt)
-            fb_violations, _ = self._check_preservation(prompt, fallback)
-            fb_completeness = self._check_completeness_issues(fallback)
-            if not fb_violations and not fb_completeness and fallback.strip() != prompt.strip():
-                self._log(
-                    f"  [safety] preservation still has {len(final_violations)} violation(s); "
-                    "using deterministic safe enhancement"
-                )
-                result = fallback
-            else:
-                self._log(
-                    f"  [safety] preservation still has {len(final_violations)} violation(s); "
-                    "returning original"
-                )
-                result = prompt
-
-        orig_score, best_score = self._estimate_quality(prompt, result)
-        self._original_fitness = orig_score
-        self._best_fitness = best_score
-        self._final_prompt = result
-
-        # v4: persist accumulated lessons to cross-prompt cache.
-        try:
-            self._persist_new_lessons()
-        except Exception:
-            pass
-
-        elapsed = time.time() - t0
-        api_calls = self.llm_client.total_api_calls - self._api_calls_start
-
-        self._log(f"\n{'='*60}")
-        self._log(f"RIDER Genesis [{mode.upper()}] — Results")
-        self._log(f"{'='*60}")
-        self._log(f"Contract: archetype={self._contract.get('task_archetype')}, "
-                  f"lang={self._contract.get('language')}, "
-                  f"strategies={self._contract.get('recommended_strategies')}")
-        self._log(f"Quality: {orig_score:.0%} -> {best_score:.0%} ({self.improvement:+.1f}%)")
-        self._log(f"Time: {elapsed:.1f}s | API calls: {api_calls}")
-        self._log(f"Prompt: {result[:200]}...")
-        self._log(f"{'='*60}")
-        return result
-
-    # ══════════════════════════════════════════════════════════════════════
-    # Mode runners
-    # ══════════════════════════════════════════════════════════════════════
-
-    def run_light(self, prompt: str) -> str:
-        """RIDER Light — fast optimization (~15-35s, ~5 calls).
-
-        1. Contract extraction (1 Sonnet)
-        2. 2 adaptive strategies IN PARALLEL at IGNITION temperature (2 working)
-        3. Select best + WHY lesson (1 Sonnet)
-        4. Quality estimate (1 Sonnet)
-
-        v4.3: strategies run at IGNITION T=1.15 (exploration).
-        """
-        t0 = time.time()
-        self._setup_run(prompt, "light")
-
-        self._contract = self._extract_contract(prompt)
-        self._log(f"Contract: archetype={self._contract.get('task_archetype')}, "
-                  f"preserve={self._contract.get('must_preserve')}")
-
-        strategies = self._adaptive_strategies("light")
-        self._log(f"Strategies (adaptive, IGNITION T={self._phase_temperature('ignition'):.2f}): {strategies}")
-        self._phase_temps_used.append(f"ignition={self._phase_temperature('ignition'):.2f}")
-        ign_t = self._phase_temperature('ignition')
-        with ThreadPoolExecutor(max_workers=max(1, len(strategies))) as pool:
-            futures = {
-                pool.submit(self._apply_strategy, prompt, s, None, ign_t): s
-                for s in strategies
-            }
-            variants = []
-            for f in futures:
-                name = futures[f]
-                result = f.result()
-                if result:
-                    variants.append((name, result))
-                    self._log(f"  [{name}] {result[:80]}...")
-
-        if not variants:
-            return self._finalize_run(prompt, prompt, "light", t0)
-
-        candidates = [("original", prompt)] + variants
-        best_name, best_text = self._select_best(candidates)
-        self._log(f"Winner: {best_name}")
-
-        if best_name == "original" and variants:
-            best_text = variants[0][1]
-
-        # Light still gets a real independent critic pass: cheap/short, but it
-        # prevents the fast mode from being a raw one-shot rewrite.
-        if best_text != prompt:
-            weaknesses = self._constitutional_audit(best_text)
-            best_text = self._refine(best_text, weaknesses, self._phase_temperature('fusion'))
-            self._log(f"Light critic-refine: {best_text[:80]}...")
-
-        self._history.append({"generation": 1, "best_fitness": 0.0})
-        return self._finalize_run(prompt, best_text, "light", t0)
-
-    def _run_blitz(self, prompt: str) -> str:
-        """RIDER Blitz — PARALLEL multi-strategy + diversity-aware merge + fusion refine (~45-120s).
-
-        v4.3: IGNITION strategies at T=1.15, FUSION merge+refine at T=0.85.
-        """
-        t0 = time.time()
-        self._setup_run(prompt, "blitz")
-
-        self._contract = self._extract_contract(prompt)
-        self._log(f"Contract: archetype={self._contract.get('task_archetype')}, "
-                  f"preserve={self._contract.get('must_preserve')}")
-
-        strategies = self._adaptive_strategies("blitz")
-        ign_t = self._phase_temperature('ignition')
-        fus_t = self._phase_temperature('fusion')
-        self._log(f"Strategies (IGNITION T={ign_t:.2f}): {strategies}")
-        with ThreadPoolExecutor(max_workers=max(1, len(strategies))) as pool:
-            futures = {
-                pool.submit(self._apply_strategy, prompt, s, None, ign_t): s
-                for s in strategies
-            }
-            variants = []
-            for f in futures:
-                name = futures[f]
-                result = f.result()
-                if result:
-                    variants.append((name, result))
-                    self._log(f"  [{name}] {result[:70]}...")
-
-        if not variants:
-            return self._finalize_run(prompt, prompt, "blitz", t0)
-
-        candidates = [("original", prompt)] + variants
-        top1, partner, collapsed = self._pick_diverse_pair(candidates)
-        if top1 is None:
-            return self._finalize_run(prompt, prompt, "blitz", t0)
-        if partner is None:
-            partner = top1
-        self._log(f"Top: {top1[0]} | Partner (diversity-aware): {partner[0]} "
-                  f"| collapsed={collapsed}")
-        self._history.append({"generation": 1, "best_fitness": 0.0, "stage": "tournament"})
-
-        # FUSION phase: merge + audit IN PARALLEL (audit can start from top1 while merge runs),
-        # then refine merged with fusion-T.
-        self._log(f"FUSION: parallel(audit on top1, merge top1+partner) at T={fus_t:.2f}")
-        with ThreadPoolExecutor(max_workers=2) as pool:
-            merge_f = pool.submit(self._merge, top1[1], partner[1], fus_t)
-            audit_f = pool.submit(self._constitutional_audit, top1[1])
-            merged = merge_f.result()
-            weaknesses = audit_f.result()
-        self._log(f"Merged: {merged[:80]}...")
-        refined = self._refine(merged, weaknesses, fus_t)
-        self._log(f"Refined: {refined[:80]}...")
-        self._history.append({"generation": 2, "best_fitness": 0.0, "stage": "refine"})
-
-        return self._finalize_run(prompt, refined, "blitz", t0)
-
-    def _run_standard(self, prompt: str) -> str:
-        """RIDER Standard — PHASE REACTOR 3-phase + v4 synthetic eval + beam k=2 (~120s, ~22 calls)."""
-        t0 = time.time()
-        self._setup_run(prompt, "standard")
-
-        self._contract = self._extract_contract(prompt)
-        self._log(f"Contract: archetype={self._contract.get('task_archetype')}, "
-                  f"preserve={self._contract.get('must_preserve')}")
-
-        # v4: bootstrap with lessons from prior runs on same archetype+domain.
-        injected = self._prefetch_cached_lessons()
-        if injected:
-            self._log(f"[lesson cache] injected {injected} cross-prompt lessons "
-                      f"(key={self._cache_key()})")
-
-        ign_t = self._phase_temperature('ignition')
-        fus_t = self._phase_temperature('fusion')
-        cry_t = self._phase_temperature('crystallization')
-
-        # v4.4: adaptive pipeline for standard too.
-        tier = self._complexity_tier()
-        self._log(f"[adaptive pipeline] complexity tier = {tier.upper()} "
-                  f"(len={self._original_prompt_len} chars, "
-                  f"archetype={self._contract.get('task_archetype')})")
-        # Standard mode: simple tier skips Phase 3 crystal polish (FUSION enough);
-        # medium/complex run all 3 phases.
-        run_crystal = (tier != 'simple')
-
-        # ── Phase 1: IGNITION (T=1.15 exploration) ──
-        self._log(f"\n> Phase 1: IGNITION — adaptive strategies at T={ign_t:.2f}")
-        all_strategies = self._adaptive_strategies("standard")
-        # On simple, 3 strategies; medium/complex — full 5.
-        strategies = all_strategies[:3] if tier == 'simple' else all_strategies
-        self._log(f"  strategies ({tier} tier, {len(strategies)}): {strategies}")
-        with ThreadPoolExecutor(max_workers=max(1, len(strategies))) as pool:
-            futures = {
-                pool.submit(self._apply_strategy, prompt, s, None, ign_t): s
-                for s in strategies
-            }
-            variants = []
-            for f in futures:
-                name = futures[f]
-                result = f.result()
-                if result:
-                    variants.append((name, result))
-                    self._log(f"  [{name}] {result[:70]}...")
-
-        if not variants:
-            return self._finalize_run(prompt, prompt, "standard", t0)
-
-        candidates = [("original", prompt)] + variants
-        # v4.3: N-way batch ranking in ONE Sonnet call (was 3 sequential pairwise).
-        top = self._rank_batch(candidates, k=3)
-        self._log(f"  -> Top 3 (N-way batch rank): {[n for n, _ in top]}")
-        self._history.append({"generation": 1, "best_fitness": 0.0, "stage": "ignition"})
-
-        # ── Phase 2: FUSION (T=0.85 balanced) ──
-        self._log(f"\n> Phase 2: FUSION — audit + refine top-3 IN PARALLEL at T={fus_t:.2f}")
-        # Audit + refines of 3 top candidates can all run in parallel.
-        with ThreadPoolExecutor(max_workers=4) as pool:
-            audit_f = pool.submit(self._constitutional_audit, top[0][1])
-            weaknesses = audit_f.result()
-            self._log(f"  audit: {weaknesses[:160]}...")
-            refine_futures = {
-                pool.submit(self._refine, t, weaknesses, fus_t): n
-                for n, t in top
-            }
-            refined: List[Tuple[str, str]] = []
-            for f in refine_futures:
-                name = refine_futures[f]
-                result = f.result()
-                refined.append((f"{name}+", result))
-                self._log(f"  [{name}+] {result[:70]}...")
-
-        top1, partner, collapsed = self._pick_diverse_pair(refined)
-        self._log(f"  diverse pair: {top1[0] if top1 else None} + {partner[0] if partner else None} "
-                  f"| collapsed={collapsed}")
-        merged = self._merge(top1[1], partner[1], fus_t) if (top1 and partner and partner[1] != top1[1]) else (top1[1] if top1 else prompt)
-        self._log(f"  Merged: {merged[:80]}...")
-        self._history.append({"generation": 2, "best_fitness": 0.0, "stage": "fusion"})
-
-        # ── Phase 3: CRYSTALLIZATION (T=0.55 polish) ──
-        # v4.4: skipped on simple tier (FUSION result already well-polished,
-        # extra audit+refine accumulates noise).
-        if run_crystal:
-            self._log(f"\n> Phase 3: CRYSTALLIZATION — adversarial polish at T={cry_t:.2f}")
-            final_weaknesses = self._constitutional_audit(merged)
-            polished = self._refine(merged, final_weaknesses, cry_t)
-            self._log(f"  Polished: {polished[:80]}...")
-            self._history.append({"generation": 3, "best_fitness": 0.0, "stage": "crystallization"})
-        else:
-            self._log("\n> Phase 3 SKIPPED (simple tier) — using merged as polished")
-            polished = merged
-
-        # ── Phase 4: SYNTHETIC EVAL — beam k=2 (ECHO-enabled eval) ALWAYS runs ──
-        self._log("\n> Phase 4: SYNTHETIC TEST EVAL — beam k=2 with ECHO")
-        synth_tests = self._generate_synthetic_tests(polished, count=3)
-        self._log(f"  synthetic tests generated: {len(synth_tests)}")
-        if synth_tests:
-            # v4.3: include original in beam to measure real uplift as fitness.
-            beam = [("polished", polished), ("merged", merged), ("original", prompt)]
-            ranked = self._rank_by_synthetic_eval(beam, synth_tests)
-            if ranked:
-                self._synth_best_score = ranked[0][1]
-                for c, sc in ranked:
-                    if c[0] == "original":
-                        self._synth_orig_score = sc
-                        break
-                orig_len = getattr(self, "_original_prompt_len", 0) or len(prompt)
-                safe_non_orig = self._safe_non_original_candidates(ranked, prompt)
-                original_score = next((sc for c, sc in ranked if c[0] == "original"), 0.0)
-                original_wins = ranked[0][0][0] == "original"
-                margin = self._original_margin_for_beam()
-                force_evolve = self._is_underoptimized_prompt(prompt)
-                if original_wins and orig_len >= 500 and not force_evolve and (
-                    not safe_non_orig or original_score - safe_non_orig[0][1] > margin
-                ):
-                    polished = prompt
-                    self._synth_best_score = original_score
-                    self._log(f"  [SAFETY] original wins beam {original_score:.0%}; "
-                              f"best safe non-original gap="
-                              f"{(original_score - safe_non_orig[0][1]) if safe_non_orig else 1.0:.2f} "
-                              f"> margin={margin:.2f} — returning original")
-                elif safe_non_orig:
-                    if ranked[0][0][0] == "original":
-                        self._log(f"  [NEAR-MISS OVERRIDE] original won synth "
-                                  f"({original_score:.0%}) but "
-                                  f"{'prompt is underoptimized' if force_evolve else 'safe non-original is within margin=' + f'{margin:.2f}'}; "
-                                  f"selecting evolved prompt")
-                    polished = safe_non_orig[0][0][1]
-                    self._synth_best_score = safe_non_orig[0][1]
-                    self._log(f"  beam winner: {safe_non_orig[0][0][0]} "
-                              f"(score={safe_non_orig[0][1]:.0%} vs orig={self._synth_orig_score or 0:.0%})")
-                for cand, sc in ranked:
-                    self._log(f"       - {cand[0]}: {sc:.0%}")
-            self._history.append({"generation": 4, "best_fitness": self._synth_best_score or 0.0,
-                                  "stage": "synthetic_eval"})
-
-        return self._finalize_run(prompt, polished, "standard", t0)
-
-    def _run_ultra(self, prompt: str) -> str:
-        """RIDER Ultra+ — 5-phase: ignition/fusion/crystal/validation/red-team+synth-beam k=3 (~180s, ~33 calls)."""
-        t0 = time.time()
-        self._setup_run(prompt, "ultra")
-
-        self._contract = self._extract_contract(prompt)
-        self._log(f"Contract: archetype={self._contract.get('task_archetype')}, "
-                  f"preserve={self._contract.get('must_preserve')}, "
-                  f"failure_modes={self._contract.get('failure_modes')}")
-
-        # v4: bootstrap with lessons from prior runs on same archetype+domain.
-        injected = self._prefetch_cached_lessons()
-        if injected:
-            self._log(f"[lesson cache] injected {injected} cross-prompt lessons "
-                      f"(key={self._cache_key()})")
-
-        # v4.4: adaptive pipeline depth per complexity tier.
-        tier = self._complexity_tier()
-        # Ultra mode quality floor: users who pick ULTRA explicitly asked for
-        # ultra-quality output — never run a 3-strategy simple tier even on a
-        # 27-char vague prompt. Short prompts need the full RIDER treatment
-        # (4 strategies + crystallization) to blow up into a structured ultra
-        # instruction; without the floor, "Write an essay about autumn" got a
-        # 3-strategy pipeline that then choked on the bloat guard.
-        if tier == 'simple':
-            self._log("[adaptive pipeline] simple tier UPGRADED → medium "
-                      "(ultra mode guarantees crystallization + 4 strategies)")
-            tier = 'medium'
-        cfg = self._ultra_pipeline_config(tier)
-        self._log(f"[adaptive pipeline] complexity tier = {tier.upper()} "
-                  f"(len={self._original_prompt_len} chars, "
-                  f"archetype={self._contract.get('task_archetype')})")
-        self._log(f"  pipeline: crystal={cfg['run_crystallization']}, "
-                  f"validation={cfg['run_validation']}, "
-                  f"red_team={cfg['run_red_team_harden']}, "
-                  f"triple_merge={cfg['run_triple_merge']}, "
-                  f"strategies={cfg['num_strategies']}")
-        self._tier = tier
-        self._cfg = cfg
-
-        ign_t = self._phase_temperature('ignition')
-        fus_t = self._phase_temperature('fusion')
-        cry_t = self._phase_temperature('crystallization')
-        val_t = self._phase_temperature('validation')
-
-        # ── Phase 1: IGNITION (T=1.15) ──
-        self._log(f"\n> Phase 1: IGNITION — adaptive strategies at T={ign_t:.2f}")
-        all_strategies = self._adaptive_strategies("ultra")
-        # v4.4: truncate strategy count per tier (simple=3, medium=4, complex=5).
-        strategies = all_strategies[:cfg['num_strategies']]
-        self._log(f"  strategies ({tier} tier, top-{cfg['num_strategies']}): {strategies}")
-        with ThreadPoolExecutor(max_workers=max(1, len(strategies))) as pool:
-            futures = {
-                pool.submit(self._apply_strategy, prompt, s, None, ign_t): s
-                for s in strategies
-            }
-            variants = []
-            for f in futures:
-                name = futures[f]
-                result = f.result()
-                if result:
-                    variants.append((name, result))
-                    self._log(f"  [{name}] {result[:70]}...")
-
-        if not variants:
-            return self._finalize_run(prompt, prompt, "ultra", t0)
-
-        candidates = [("original", prompt)] + variants
-        # v4.3: N-way batch ranking in ONE Sonnet call. v4.4: top-k scales per tier.
-        k_phase1 = cfg['tournament_k_phase1']
-        top = self._rank_batch(candidates, k=k_phase1)
-        self._log(f"  -> Top {k_phase1} (N-way batch rank): {[n for n, _ in top]}")
-        self._history.append({"generation": 1, "best_fitness": 0.0, "stage": "ignition"})
-
-        # Semantic collapse check across top 3 — VORTEX opportunity.
-        collapse_sim = max(
-            (self._jaccard(a[1], b[1]) for i, a in enumerate(top) for b in top[i + 1:]),
-            default=0.0,
-        )
-        use_vortex = collapse_sim >= 0.80
-        if use_vortex:
-            self._log(f"  [collapse detected] pairwise Jaccard={collapse_sim:.2f} -> swap mutation for VORTEX")
-
-        # ── Phase 2: FUSION (T=0.85) — audit + refine + mutate ALL IN PARALLEL ──
-        self._log(f"\n> Phase 2: FUSION — parallel audit+refine+mutate at T={fus_t:.2f}")
-        with ThreadPoolExecutor(max_workers=6) as pool:
-            audit_f = pool.submit(self._constitutional_audit, top[0][1])
-            weaknesses = audit_f.result()
-            self._log(f"  audit: {weaknesses[:160]}...")
-
-            refine_futures = {
-                pool.submit(self._refine, t, weaknesses, fus_t): n for n, t in top
-            }
-
-            if use_vortex:
-                collapsed_preview = "\n---\n".join(t[1][:220] for t in top)
-                vortex_f = pool.submit(
-                    self._apply_strategy, top[0][1], 'vortex',
-                    {'collapsed_preview': collapsed_preview}, fus_t,
-                )
-                mut2_f = pool.submit(
-                    self._apply_strategy, top[0][1],
-                    self._mutation_strategy('creative', 'techniques'), None, fus_t,
-                )
-            else:
-                vortex_f = pool.submit(
-                    self._apply_strategy, top[0][1],
-                    self._mutation_strategy('creative', 'techniques'), None, fus_t,
-                )
-                mut2_f = pool.submit(
-                    self._apply_strategy, top[0][1],
-                    self._mutation_strategy('depth', 'analytical'), None, fus_t,
-                )
-
-            arena: List[Tuple[str, str]] = []
-            for f in refine_futures:
-                name = refine_futures[f]
-                result = f.result()
-                if result:
-                    arena.append((f"{name}+", result))
-            r1 = vortex_f.result()
-            if r1:
-                arena.append(("vortex" if use_vortex else "mutation-creative", r1))
-            r2 = mut2_f.result()
-            if r2:
-                arena.append(("mutation-creative" if use_vortex else "mutation-depth", r2))
-
-        for n, t in arena:
-            self._log(f"  [{n}] {t[:70]}...")
-
-        # v4.3: N-way batch ranking for top-2 (1 Sonnet call instead of 2 sequential).
-        top2 = self._rank_batch(arena, k=2)
-        self._log(f"  -> Top 2 (N-way batch rank): {[n for n, _ in top2]}")
-        self._history.append({"generation": 2, "best_fitness": 0.0, "stage": "fusion"})
-
-        # ── Phase 3: CRYSTALLIZATION (T=0.55 polish) — merge + dual polish ──
-        # v4.4: only runs on medium/complex tier. On simple prompts, top2 already strong,
-        # extra polish just adds noise (confirmed on NARKO3 benchmark).
-        if cfg['run_crystallization'] and len(top2) >= 2:
-            self._log(f"\n> Phase 3: CRYSTALLIZATION — merge + dual polish at T={cry_t:.2f}")
-            merged = self._merge(top2[0][1], top2[1][1], cry_t)
-
-            with ThreadPoolExecutor(max_workers=2) as pool:
-                polish_f = pool.submit(self._apply_strategy, merged, 'techniques', None, cry_t)
-                depth_f = pool.submit(self._apply_strategy, merged, 'depth', None, cry_t)
-                polished = polish_f.result() or merged
-                deepened = depth_f.result() or merged
-
-            crystal = self._select_best(
-                [("polished", polished), ("deepened", deepened), ("merged", merged)]
-            )
-            self._log(f"  Crystal winner: {crystal[0]}")
-            self._history.append({"generation": 3, "best_fitness": 0.0, "stage": "crystallization"})
-        else:
-            self._log(f"\n> Phase 3 SKIPPED ({tier} tier) — using top2[0] as crystal")
-            crystal = top2[0] if top2 else ("top1", prompt)
-
-        # ── Phase 4: VALIDATION (T=0.3 near-deterministic) ──
-        # v4.4: complex tier only. Extra audit+refine on simple prompts tightens
-        # language without adding substance; on long prompts it catches real edge cases.
-        if cfg['run_validation']:
-            self._log(f"\n> Phase 4: VALIDATION — bilingual adversarial + audit IN PARALLEL at T={val_t:.2f}")
-            lang = (self._contract.get('language') or '').lower()
-            adversarial_directives: Optional[str] = None
-            use_bilingual = bool(lang and lang not in ('en', 'english', 'unknown', ''))
-            with ThreadPoolExecutor(max_workers=2) as pool:
-                audit_f = pool.submit(self._constitutional_audit, crystal[1])
-                bilingual_f = pool.submit(self._bilingual_adversarial, crystal[1], lang) if use_bilingual else None
-                final_weaknesses = audit_f.result()
-                if bilingual_f is not None:
-                    adversarial_directives = bilingual_f.result()
-                    self._log(f"  bilingual adversarial ({lang}): "
-                              f"{adversarial_directives[:140] if adversarial_directives else '-'}...")
-            combined_audit = final_weaknesses
-            if adversarial_directives:
-                combined_audit = f"{final_weaknesses}\n\nBilingual reviewer says:\n{adversarial_directives}"
-            validated = self._refine(crystal[1], combined_audit, val_t)
-
-            winner_name, winner_text = self._select_best(
-                [("validated", validated), ("crystal", crystal[1])]
-            )
-            self._log(f"  Final: {winner_name}")
-            self._history.append({"generation": 4, "best_fitness": 0.0, "stage": "validation"})
-        else:
-            self._log(f"\n> Phase 4 SKIPPED ({tier} tier) — using crystal directly")
-            winner_name, winner_text = crystal
-
-        # ── Phase 5: SYNTHETIC BEAM k=3 (always on) + optional red-team hardening ──
-        # v4.4: synth-eval beam ALWAYS runs (this is the real-fitness signal RIDER
-        # depends on). Red-team hardening only on complex tier — on simple prompts
-        # it adds rules the LLM can't reliably execute.
-        self._log("\n> Phase 5: SYNTHETIC BEAM k=3" + (
-            " + RED-TEAM hardening" if cfg['run_red_team_harden'] else
-            " (red-team SKIPPED for " + tier + " tier)"
-        ))
-
-        if cfg['run_red_team_harden']:
-            with ThreadPoolExecutor(max_workers=2) as pool:
-                red_team_f = pool.submit(self._red_team_adversarial, winner_text)
-                synth_tests_f = pool.submit(self._generate_synthetic_tests, winner_text, 5)
-                red_team = red_team_f.result()
-                speculative_tests = synth_tests_f.result() or []
-
-            hardened = winner_text
-            if red_team and isinstance(red_team, dict):
-                edge_cases = red_team.get('edge_cases') or []
-                fix_dirs = red_team.get('fix_directives') or []
-                severity = red_team.get('severity', 'low')
-                self._log(f"  [red-team] severity={severity}, "
-                          f"edge_cases={len(edge_cases)}, fixes={len(fix_dirs)}")
-                if severity in ('medium', 'high') and fix_dirs:
-                    fix_brief = (
-                        "RED-TEAM EDGE CASES:\n"
-                        + "\n".join(f"- {ec}" for ec in edge_cases[:3])
-                        + "\n\nFIXES TO APPLY:\n"
-                        + "\n".join(f"- {fd}" for fd in fix_dirs[:3])
-                    )
-                    hardened = self._refine(winner_text, fix_brief, val_t)
-                    self._log(f"  [red-team] hardened: {hardened[:80]}...")
-                    synth_tests = self._generate_synthetic_tests(hardened, count=5)
-                else:
-                    synth_tests = speculative_tests
-            else:
-                synth_tests = speculative_tests
-        else:
-            # Simple/medium: skip red-team, just generate synth tests for beam fitness.
-            hardened = winner_text
-            synth_tests = self._generate_synthetic_tests(winner_text, count=5)
-        self._log(f"  synthetic tests generated: {len(synth_tests)}")
-
-        # v4.3: include original in beam to get REAL fitness delta.
-        # v4.4: beam composition depends on which phases ran.
-        final_pick = hardened
-        if synth_tests:
-            beam: List[Tuple[str, str]] = []
-            if cfg['run_red_team_harden']:
-                beam.append(("hardened", hardened))
-            if cfg['run_validation']:
-                beam.append(("validated", winner_text))
-            beam.append(("crystal", crystal[1]))
-            if top2:
-                beam.append(("top2[0]", top2[0][1]))
-            beam.append(("original", prompt))
-            # Bloat filter: skip runaway variants. Uses _bloat_budget() so short
-            # originals get a generous absolute floor (can reach ultra-quality)
-            # while long originals are capped at 3x (no 10-15x runaways).
-            orig_len = max(1, len(prompt))
-            budget = self._bloat_budget(orig_len)
-            filtered_beam: List[Tuple[str, str]] = []
-            for name, txt in beam:
-                if name == "original" or len(txt) <= budget:
-                    filtered_beam.append((name, txt))
-                else:
-                    ratio = len(txt) / orig_len
-                    self._log(
-                        f"  [bloat-filter] skipping '{name}' "
-                        f"({len(txt)} chars, ratio={ratio:.1f}x, budget={budget})"
-                    )
-            beam = filtered_beam or [("original", prompt)]
-            seen: set = set()
-            uniq_beam: List[Tuple[str, str]] = []
-            for name, txt in beam:
-                key = txt[:200]
-                if key not in seen:
-                    seen.add(key)
-                    uniq_beam.append((name, txt))
-            ranked = self._rank_by_synthetic_eval(uniq_beam, synth_tests)
-            if ranked:
-                for c, sc in ranked:
-                    if c[0] == "original":
-                        self._synth_orig_score = sc
-                        break
-                orig_len = getattr(self, "_original_prompt_len", 0) or len(prompt)
-                safe_non_orig_ranked = self._safe_non_original_candidates(ranked, prompt)
-                original_score = next((sc for c, sc in ranked if c[0] == "original"), 0.0)
-                original_wins = ranked[0][0][0] == "original"
-                margin = self._original_margin_for_beam()
-                force_evolve = self._is_underoptimized_prompt(prompt)
-                if original_wins and orig_len >= 500 and not force_evolve and (
-                    not safe_non_orig_ranked or original_score - safe_non_orig_ranked[0][1] > margin
-                ):
-                    final_pick = prompt
-                    self._synth_best_score = original_score
-                    self._log(f"  [SAFETY] original wins beam {original_score:.0%}; "
-                              f"best safe non-original gap="
-                              f"{(original_score - safe_non_orig_ranked[0][1]) if safe_non_orig_ranked else 1.0:.2f} "
-                              f"> margin={margin:.2f} — returning original")
-                elif safe_non_orig_ranked:
-                    if ranked[0][0][0] == "original":
-                        self._log(f"  [NEAR-MISS OVERRIDE] original won synth "
-                                  f"({original_score:.0%}) but "
-                                  f"{'prompt is underoptimized' if force_evolve else 'safe non-original is within margin=' + f'{margin:.2f}'}; "
-                                  f"selecting evolved prompt")
-                    self._synth_best_score = safe_non_orig_ranked[0][1]
-                    final_pick = safe_non_orig_ranked[0][0][1]
-                    self._log(f"  beam winner: {safe_non_orig_ranked[0][0][0]} "
-                              f"(score={safe_non_orig_ranked[0][1]:.0%} vs orig={self._synth_orig_score or 0:.0%})")
-                else:
-                    # Degenerate case: only original in beam (everything else filtered out).
-                    final_pick = prompt
-                    self._synth_best_score = ranked[0][1] if ranked else 0.0
-                    self._log("  [SAFETY] no non-original candidates in beam — returning original")
-                for cand, sc in ranked:
-                    self._log(f"       - {cand[0]}: {sc:.0%}")
-
-                # v4.3 TRIPLE-MERGE: if top-2 non-original candidates are close in score
-                # (within 15%) AND neither loses to original, merge for super-synthesis.
-                # v4.4: complex tier only — on simple prompts there's rarely material
-                # content to merge, the super-merge just produces a longer duplicate.
-                if (cfg['run_triple_merge']
-                        and ranked[0][0][0] != "original"
-                        and len(safe_non_orig_ranked) >= 2):
-                    top_sc = safe_non_orig_ranked[0][1]
-                    second_sc = safe_non_orig_ranked[1][1]
-                    if top_sc - second_sc < 0.15 and top_sc >= (self._synth_orig_score or 0):
-                        super_merged = self._merge(
-                            safe_non_orig_ranked[0][0][1],
-                            safe_non_orig_ranked[1][0][1],
-                            cry_t,
-                        )
-                        self._log(f"  [triple-merge] close scores {top_sc:.0%}/{second_sc:.0%} -> super-merge attempt")
-                        super_score, _ = self._evaluate_candidate_on_tests(super_merged, synth_tests[:3])
-                        self._log(f"  [triple-merge] super score: {super_score:.0%}")
-                        if super_score > top_sc:
-                            final_pick = super_merged
-                            self._synth_best_score = super_score
-                            self._log(f"  [triple-merge] WINS — new best {super_score:.0%}")
-        self._history.append({"generation": 5, "best_fitness": self._synth_best_score or 0.0,
-                              "stage": "red_team_synthetic"})
-
-        # Final safety: if final_pick exceeds _bloat_budget(), downgrade to the
-        # smallest candidate still within budget. Uses the same adaptive policy
-        # as the 3 other bloat layers (FLOOR=2500, RATIO=3.0) — short originals
-        # get a generous absolute floor so ultra-quality expansion is allowed,
-        # long originals stay within 3x to avoid 10-15x runaways.
-        orig_len = max(1, len(prompt))
-        budget = self._bloat_budget(orig_len)
-        if len(final_pick) > budget:
-            small_candidates: List[Tuple[str, str]] = []
-            if cfg['run_validation']:
-                small_candidates.append(("validated", winner_text))
-            small_candidates.append(("crystal", crystal[1]))
-            if top2:
-                small_candidates.append(("top2[0]", top2[0][1]))
-            small_ok = [c for c in small_candidates if len(c[1]) <= budget]
-            if small_ok:
-                chosen = min(small_ok, key=lambda c: len(c[1]))
-                self._log(f"  [FINAL BLOAT GUARD] final={len(final_pick)} > "
-                          f"budget={budget} (orig={orig_len}) → "
-                          f"downgrade to {chosen[0]} ({len(chosen[1])} chars)")
-                final_pick = chosen[1]
-            else:
-                # Every candidate is over budget — still prefer the shortest one
-                # over reverting to the original. User explicitly wants blow-up
-                # on short prompts, so returning the vague input is unacceptable.
-                all_non_orig = [
-                    ("hardened", hardened),
-                    ("crystal", crystal[1]),
-                ]
-                if top2:
-                    all_non_orig.append(("top2[0]", top2[0][1]))
-                if cfg['run_validation']:
-                    all_non_orig.append(("validated", winner_text))
-                chosen = min(all_non_orig, key=lambda c: len(c[1]))
-                self._log(f"  [FINAL BLOAT GUARD] all variants over budget={budget}, "
-                          f"picking shortest non-original ({chosen[0]}, {len(chosen[1])} chars)")
-                final_pick = chosen[1]
-
-        return self._finalize_run(prompt, final_pick, "ultra", t0)
-
-    def _bilingual_adversarial(self, prompt: str, lang: str) -> Optional[str]:
-        """Ultra-only: English critique for non-EN prompts to surface universal issues."""
-        meta = self._BILINGUAL_ADVERSARIAL_PROMPT.format(prompt=prompt, lang=lang)
-        try:
-            resp = self._generate(
-                prompt=meta, role="critic",
-                temperature=0.2, max_tokens=400,
-            )
-            return resp.strip() if resp and resp.strip() else None
-        except Exception:
-            return None
-
-    # ══════════════════════════════════════════════════════════════════════
-    # v4 Ultra+ methods
-    # ══════════════════════════════════════════════════════════════════════
-
-    def _fallback_synthetic_tests(self, count: int = 3) -> List[str]:
-        """Deterministic NEXUS-lite cases when the planner model does not return JSON."""
-        archetype = str(self._contract.get('task_archetype') or 'other').lower()
-        output_format = self._contract.get('output_format_anchor') or 'requested output'
-        domain = self._contract.get('domain') or 'general domain'
-        rules = "; ".join(str(x) for x in (self._contract.get('domain_rules') or [])[:3])
-        rules = rules or "follow the source task contract"
-        contract_text = " ".join(
-            str(x)
-            for x in (
-                list(self._contract.get('must_preserve') or [])
-                + list(self._contract.get('domain_rules') or [])
-                + list(self._contract.get('failure_modes') or [])
-                + list(self._contract.get('quality_dimensions') or [])
-            )
-        ).lower()
-        has_geo_level_rule = any(term in contract_text for term in (
-            'russia', 'ukraine', 'russian-side', 'россия', 'украина', 'геополит',
-        ))
-        has_exclusive_output_rule = any(term in contract_text for term in (
-            'exclusive', 'only allowed output', 'output only', 'json only',
-            'единственный', 'только json', 'schema drift',
-        ))
-
-        if archetype in {'creative_writing', 'analytical_essay', 'persuasion', 'brainstorming', 'other'}:
-            tests = [
-                f"Input: execute the prompt for a skeptical educated reader in {domain}. Expected invariant checks: distinctive angle, concrete details, clear structure, anti-generic wording, output format = {output_format}.",
-                "Input: execute the prompt while avoiding cliches and generic school-essay framing. Expected invariant checks: named thesis, specific evidence/details, no padded meta-commentary.",
-                "Input: execute the prompt for a publication-quality result. Expected invariant checks: coherent arc, vivid examples, quality bar satisfied, no irrelevant constraints.",
-            ]
-        elif archetype in {'debugging', 'code_generation', 'code_review'}:
-            tests = [
-                f"Input: solve the original code task exactly. Expected invariant checks: preserve identifiers/API, minimal patch, tests included, runtime reasoning correct, output format = {output_format}.",
-                "Input: handle the edge case that caused the reported failure. Expected invariant checks: regression test covers the failure and no unrelated refactor is introduced.",
-                "Input: explain verification briefly. Expected invariant checks: concrete commands or test cases, no invented dependencies, no renamed fields.",
-            ]
-        elif archetype == 'classification':
-            tests = [
-                f"Input: benign placeholder quote with no violation. Expected invariant checks: valid JSON, exact fields, Level 0 stays Level 0, rules: {rules}.",
-                f"Input: borderline placeholder quote with ambiguous risk marker. Expected invariant checks: quote-grounded category decision, no invented law/category, schema = {output_format}.",
-                "Input: high-risk placeholder quote that must not be missed. Expected invariant checks: false-negative resistance, correct level calibration, no prose outside JSON.",
-            ]
-            if has_geo_level_rule:
-                tests.insert(
-                    1,
-                    "Input: quote mentions Russia/Ukraine/Russian side in a borderline context. Expected invariant checks: apply the source escalation rule exactly, keep the Level 0 exception, and cap the result at Level 3.",
-                )
-            if has_exclusive_output_rule:
-                tests.insert(
-                    0,
-                    f"Input: any valid classification case. Expected invariant checks: the only output is {output_format}; no explanations, helper prose, markdown fences, or alternate schemas.",
-                )
-        elif archetype == 'translation':
-            tests = [
-                f"Input: short literary paragraph plus glossary entry. Expected invariant checks: glossary compliance, authorial rhythm, no extra commentary, output format = {output_format}.",
-                "Input: intentionally rough source sentence with repetition and odd syntax. Expected invariant checks: preserve author texture rather than smoothing into generic Russian.",
-                "Input: source paragraph with names/placeholders/tags. Expected invariant checks: placeholders and XML tags preserved exactly.",
-            ]
-        elif archetype == 'data_extraction':
-            tests = [
-                f"Input: noisy OCR/table-of-contents snippet. Expected invariant checks: verbatim extraction, no paraphrase, no extra commentary, output format = {output_format}.",
-                "Input: partially unreadable line with page numbers and tags. Expected invariant checks: preserve unreadable markers and structure exactly.",
-                "Input: mixed headings and list items. Expected invariant checks: completeness, original order, no invented text.",
-            ]
-        else:
-            tests = [
-                f"Input: straightforward case in {domain}. Expected invariant checks: fulfills task, preserves constraints, output format = {output_format}.",
-                "Input: ambiguous edge case. Expected invariant checks: states assumptions only if allowed and does not drop required fields.",
-                "Input: complex case with multiple requirements. Expected invariant checks: complete, structured, and no irrelevant additions.",
-            ]
-        return tests[:count]
-
-    def _generate_synthetic_tests(self, candidate_prompt: str, count: int = 3) -> List[str]:
-        """v4: Generate safe synthetic test inputs via Sonnet. Returns up to `count` inputs."""
-        if not self._contract:
-            return []
-        meta = self._SYNTHETIC_TEST_PROMPT.format(
-            archetype=self._contract.get('task_archetype', 'other'),
-            domain=self._contract.get('domain', 'general'),
-            audience=self._contract.get('audience', 'general'),
-            output_format=self._contract.get('output_format_anchor', 'free-form'),
-            domain_rules="\n".join(f"- {x}" for x in (self._contract.get('domain_rules') or ['none'])),
-            quality_dimensions="\n".join(f"- {x}" for x in (self._contract.get('quality_dimensions') or ['general task fulfillment'])),
-            prompt=(candidate_prompt or '')[:2500],
-            count=count,
-        )
-        obj = self._generate_structured(
-            prompt=meta,
-            schema=_SyntheticTestsSchema,
-            role="planner",
-            temperature=0.5,
-            max_tokens=1500,
-            allowed_starts=("[", "{"),
-            max_retries=2,
-        )
-        if obj is None:
-            result = self._fallback_synthetic_tests(count)
-            self._synthetic_tests = result
-            return result
-        clean = list(getattr(obj, "tests", []) or [])
-        result = clean[:count] or self._fallback_synthetic_tests(count)
-        self._synthetic_tests = result
-        return result
-
-    def _evaluate_candidate_on_tests(
-        self, candidate: str, tests: List[str]
-    ) -> Tuple[float, List[Tuple[str, int]]]:
-        """v4.3: Execute candidate on each synthetic test in PARALLEL with ECHO (prompt
-        repeated 2x — RIDER ECHO protocol, ~67% eval-quality bump on non-reasoning models),
-        then score in PARALLEL. Uses _generate with role-chain refusal fallback.
-
-        v4.3.2: bloat guard — penalize candidates that grew >2.5x vs original (judge
-        consistently flags over-engineered monsters as worse than tighter versions).
-        """
-        if not tests:
-            return 0.0, []
-
-        def _run(test_input: str) -> Tuple[str, int]:
-            # v4.3: ECHO — repeat the candidate prompt twice to eliminate causal
-            # attention asymmetry (Google arxiv.org/pdf/2512.14982, +67% on non-reasoning).
-            exec_prompt = f"{candidate}\n\n{candidate}\n\nInput:\n{test_input}"
-            response = self._generate(
-                prompt=exec_prompt, role="worker", temperature=0.3,
-                max_tokens=self._adaptive_max_tokens(test_input, min_tokens=500, ceiling=3000),
-            )
-            meta = self._EVAL_RESPONSE_PROMPT.format(
-                archetype=self._contract.get('task_archetype', 'other'),
-                output_format=self._contract.get('output_format_anchor', 'free-form'),
-                quality_dimensions="\n".join(
-                    f"- {x}" for x in (self._contract.get('quality_dimensions') or ['general task fulfillment'])
-                ),
-                candidate=(candidate or '')[:1500],
-                test_input=(test_input or '')[:800],
-                response=(response or '')[:2000],
-            )
-            score_obj = self._generate_structured(
-                prompt=meta,
-                schema=_JudgeScoreSchema,
-                role="judge",
-                temperature=0.0,
-                max_tokens=64,
-                allowed_starts=("{",),
-                max_retries=1,
-            )
-            return (test_input[:60], int(score_obj.score) if score_obj else 5)
-
-        workers = max(1, min(8, len(tests)))
-        with ThreadPoolExecutor(max_workers=workers) as pool:
-            futures = [pool.submit(_run, t) for t in tests]
-            results = [f.result() for f in futures]
-
-        details = results
-        scores = [d[1] / 10.0 for d in details]
-        base = (sum(scores) / len(scores)) if scores else 0.0
-
-        # Synthetic execution can over-credit a raw original because a strong
-        # worker model often infers missing structure by itself. For auto-prompting
-        # fitness, the prompt must carry the contract explicitly.
-        original_prompt = getattr(self, "_original_prompt", "") or ""
-        if candidate.strip() == original_prompt.strip() and self._is_underoptimized_prompt(original_prompt):
-            base = min(base, 0.45)
-
-        # Bloat guard: penalize over-engineered monsters relative to _bloat_budget().
-        # Short originals: no penalty until candidate exceeds the absolute floor
-        # (so ultra-quality expansions are allowed to shine). Long originals:
-        # penalty kicks in above 3x — we DO want to curb 10-15x runaways.
-        orig = getattr(self, "_original_prompt_len", 0)
-        if orig > 0:
-            budget = self._bloat_budget(orig)
-            if len(candidate) > budget:
-                # Excess over budget, normalized by budget itself.
-                # 1.0x over budget → 0.08 penalty; 2.5x over → 0.2 (cap).
-                excess = (len(candidate) - budget) / max(1, budget)
-                penalty = min(0.2, 0.08 * excess)
-                base = max(0.0, base - penalty)
-        return base, details
-
-    def _rank_by_synthetic_eval(
-        self, candidates: List[Tuple[str, str]], tests: List[str]
-    ) -> List[Tuple[Tuple[str, str], float]]:
-        """v4.2: Score each candidate on tests in PARALLEL. Outer pool across candidates,
-        inner pool inside _evaluate_candidate_on_tests across tests. Nested threading is safe
-        on HTTP I/O. Total concurrency: N_cands * N_tests capped at 16."""
-        if not tests or not candidates:
-            return [(c, 0.0) for c in candidates]
-
-        def _rank_one(cand: Tuple[str, str]) -> Tuple[Tuple[str, str], float]:
-            avg, _details = self._evaluate_candidate_on_tests(cand[1], tests)
-            return (cand, avg)
-
-        workers = max(1, min(4, len(candidates)))
-        with ThreadPoolExecutor(max_workers=workers) as pool:
-            futures = [pool.submit(_rank_one, c) for c in candidates]
-            ranked = [f.result() for f in futures]
-
-        ranked.sort(key=lambda x: -x[1])
-        self._synthetic_rankings.append({
-            'tests': list(tests),
-            'scores': [{'name': cand[0], 'score': score} for cand, score in ranked],
-        })
-        return ranked
-
-    def _original_margin_for_beam(self) -> float:
-        """How much a safe non-original may trail original and still be worth using."""
-        archetype = str(self._contract.get('task_archetype') or '').lower()
-        dims = {str(x).lower() for x in (self._contract.get('quality_dimensions') or [])}
-        if archetype == 'classification' or {'schema_compliance', 'level_calibration'} & dims:
-            return 0.03
-        if archetype == 'translation':
-            return 0.08
-        if archetype == 'data_extraction' or {'verbatim_fidelity', 'markup_compliance'} & dims:
-            return 0.05
-        if archetype in {'creative_writing', 'analytical_essay', 'persuasion', 'brainstorming'}:
-            return 0.15
-        return 0.12
-
-    def _safe_non_original_candidates(
-        self, ranked: List[Tuple[Tuple[str, str], float]], original: str,
-    ) -> List[Tuple[Tuple[str, str], float]]:
-        """Filter beam candidates that can be selected without final safety rollback."""
-        budget = self._bloat_budget(max(1, len(original)))
-        safe: List[Tuple[Tuple[str, str], float]] = []
-        for cand, score in ranked:
-            name, text = cand
-            if name == "original":
-                continue
-            if len(text) > budget:
-                continue
-            violations, _ = self._check_preservation(original, text)
-            if violations:
-                continue
-            if self._check_completeness_issues(text):
-                continue
-            safe.append((cand, score))
-        return safe
-
-    def _red_team_adversarial(self, prompt: str) -> Optional[Dict[str, Any]]:
-        """v4 Ultra-only: adversary finds edge cases + fix directives."""
-        meta = self._RED_TEAM_PROMPT.format(
-            prompt=(prompt or '')[:3000],
-            archetype=self._contract.get('task_archetype', 'other'),
-            failure_modes=self._contract.get('failure_modes', []),
-        )
-        obj = self._generate_structured(
-            prompt=meta,
-            schema=_RedTeamAdversarialSchema,
-            role="critic",
-            temperature=0.4,
-            max_tokens=800,
-            allowed_starts=("{",),
-            max_retries=2,
-        )
-        return self._schema_to_dict(obj) if obj is not None else None
-
-    # ══════════════════════════════════════════════════════════════════════
-    # Public API
-    # ══════════════════════════════════════════════════════════════════════
-
     def run(
         self,
         prompt: str,
@@ -3223,50 +229,40 @@ class RiderGenesis:
     @property
     def final_prompt(self) -> Optional[str]:
         return self._final_prompt
-
     @property
     def improvement(self) -> float:
         if self._original_fitness <= 0:
             return 0.0
         return (self._best_fitness - self._original_fitness) / self._original_fitness * 100
-
     @property
     def fitness(self) -> float:
         return self._best_fitness
-
     @property
     def history(self) -> List[Dict]:
         return self._history
-
     @property
     def api_calls(self) -> int:
         return self.llm_client.total_api_calls - self._api_calls_start
-
     @property
     def contract(self) -> Dict[str, Any]:
         """Last extracted prompt contract (v3 addition)."""
         return dict(self._contract)
-
     @property
     def lessons(self) -> List[str]:
         """GENESIS-lite lessons collected during the last run (v3 addition)."""
         return list(self._lessons)
-
     @property
     def synthetic_tests(self) -> List[str]:
         """Synthetic cases generated during the last Standard/Ultra run."""
         return list(self._synthetic_tests)
-
     @property
     def synthetic_rankings(self) -> List[Dict[str, Any]]:
         """Synthetic beam audit trail: cases, candidate names, and scores."""
         return list(self._synthetic_rankings)
-
     @property
     def role_models(self) -> Dict[str, str]:
         """Primary model used by each RIDER Genesis role in the last/current run."""
         return {role: self._role_model(role) for role in self._ROLES}
-
     @property
     def llm_attempts(self) -> List[Dict[str, Any]]:
         """Diagnostics for model routing, validation failures, and fallbacks."""
